% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockSamplesSIRfsMCMC.R
\name{adaptiveSIRfsMCMC7_blockedIS}
\alias{adaptiveSIRfsMCMC7_blockedIS}
\title{Calculate the amount of blockSamples which need to be made
To accomodate for a potentially different final block size (blockSamples is not integer), calculate the residual noDraws
Create Storage Matrix
Proposal Acceptance Counter
Calculate the amount of blockSamples which need to be made
To accomodate for a potentially different final block size (blockSamples is not integer), calculate the residual noDraws
Store State
Calculating Summary Statistics for samples}
\usage{
adaptiveSIRfsMCMC7_blockedIS(
  obsTransData,
  I_0,
  obsTimes,
  N,
  beta0,
  gamma0,
  thetaLim,
  lambda0,
  V,
  firstRun = TRUE,
  noDraws = 2 * N - I_0,
  blockSize,
  noIts,
  burnIn = 0,
  lagMax = NA,
  thinningFactor = 1,
  delta
)
}
\description{
Calculate the amount of blockSamples which need to be made
To accomodate for a potentially different final block size (blockSamples is not integer), calculate the residual noDraws
Create Storage Matrix
Proposal Acceptance Counter
Calculate the amount of blockSamples which need to be made
To accomodate for a potentially different final block size (blockSamples is not integer), calculate the residual noDraws
Store State
Calculating Summary Statistics for samples
}
