% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaptiveSIS_PM-MCMC.R
\name{adaptiveSIS_PseudoMarginalMCMC}
\alias{adaptiveSIS_PseudoMarginalMCMC}
\title{Adaptive Pseudo-Marginal MCMC for SIS Epidemic Panel Data}
\usage{
adaptiveSIS_PseudoMarginalMCMC(
  obsTransData,
  I_0,
  obsTimes,
  N,
  beta0,
  gamma0,
  lambda0 = 2.38/sqrt(2),
  V0 = diag(c(1/N, 1)),
  noSims,
  noIts,
  burnIn,
  lagMax = NA,
  thinningFactor = 1,
  parallel = FALSE,
  noCores,
  delta = 0.05
)
}
\arguments{
\item{obsTransData}{Interpanel transition data.}

\item{I_0}{Initial number of infectives in the population.}

\item{obsTimes}{Times at which epidemic cohort were followed up.}

\item{N}{Population size.}

\item{beta0}{Starting value for infectious process parameter.}

\item{gamma0}{Starting value for removal/recovery process parameter.}

\item{lambda0}{Starting value for RWM proposal parameter which is to be adapted.}

\item{V0}{Starting state for RWM proposal Covariance matrix which is to be adapted.}

\item{noSims}{Number of epidemic simulations to be used per iteration for estimation of likelihood.}

\item{noIts}{Number of MCMC iterations.}

\item{lagMax}{Plotting parameter for acf() function.}

\item{thinningFactor}{Controls the factor by which MCMC samples are thinned, to reduce dependency.}

\item{parallel}{Are epidemic simulations run in parallel?}

\item{noCores}{If epidemic simulations are run in parallel, this is the number of cores utilised.}

\item{delta}{Probability that a proposal is made based on the starting proposal conditions.}
}
\value{
Proposal parameters which can be used for more optimal exploration of target distribution (plus MCMC summary).
Create Storage Matrix
Proposal Acceptance Counter
Propose new beta and gamma using Multiplicative RW propsal
Store State
Calculating Summary Statistics for samples
}
\description{
Adapts proposal parameters for with a view of optimal of a target using Pseudo-Marginal MCMC scheme.
}
\seealso{
Other Panel Data MCMC: 
\code{\link{SIR_fsMCMC_blockedIS}()},
\code{\link{adaptiveSIR_PseudoMarginalMCMC}()},
\code{\link{adaptiveSISfsMCMC}()}
}
\concept{Panel Data MCMC}
