% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptiveSISfsMCMC.R
\name{adaptiveSISfsMCMC}
\alias{adaptiveSISfsMCMC}
\title{Adaptive fsMCMC MCMC for SIS Epidemic Panel Data}
\usage{
adaptiveSISfsMCMC(
  obsTransData,
  I_0,
  obsTimes,
  N,
  beta0,
  gamma0,
  thetaLim,
  lambda0,
  V0,
  delta,
  noDraws,
  s,
  noIts,
  burnIn = 0,
  lagMax = NA,
  thinningFactor = 1
)
}
\arguments{
\item{obsTransData}{Interpanel transition data.}

\item{I_0}{Initial number of infectives in the population.}

\item{obsTimes}{Times at which epidemic cohort were followed up.}

\item{N}{Population size.}

\item{beta0}{Starting value for infectious process parameter.}

\item{gamma0}{Starting value for removal/recovery process parameter.}

\item{lambda0}{Starting value for RWM proposal parameter which is to be adapted.}

\item{V0}{Starting state for RWM proposal Covariance matrix which is to be adapted.}

\item{delta}{Probability that a proposal is made based on the starting proposal conditions.}

\item{noDraws}{Number of underlying variables to be drawn (i.e max. number of events in observation period)}

\item{s}{Number of underlying random variables of the process to refresh.}

\item{noIts}{Number of MCMC iterations.}

\item{lagMax}{Plotting parameter for acf() function.}

\item{thinningFactor}{Controls the factor by which MCMC samples are thinned, to reduce dependency.}
}
\value{
Proposal parameters which can be used for more optimal exploration of target distribution (plus MCMC summary).
}
\description{
Adapts proposal parameters for with a view of optimal of a target using forward simulation MCMC scheme.
}
\seealso{
Other Panel Data MCMC: 
\code{\link{SIR_fsMCMC_blockedIS}()},
\code{\link{adaptiveSIR_PseudoMarginalMCMC}()},
\code{\link{adaptiveSIS_PseudoMarginalMCMC}()}
}
\concept{Panel Data MCMC}
