% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.ts.r
\name{agg.ts}
\alias{agg.ts}
\title{Change time scale}
\usage{
agg.ts(ts, len)
}
\arguments{
\item{ts}{response vector of numeric values representing the time series to be modeled. By default this is set to 'g.ts' the global time series which MUST be specified by the user before calling this function}

\item{len}{numeric value indicating vector length}
}
\description{
Takes a time series and collapses to different temporal units. See serf.fit function for example on how to fit model.
}

