% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{as.intervention}
\alias{as.intervention}
\title{Convert a list to a intervention object}
\usage{
as.intervention(x)
}
\arguments{
\item{x}{A list, or an object that inherits from a list.}
}
\value{
A \link{intervention} class object.
}
\description{
Convert a list to a intervention object
}
\examples{
# prepare a list
npi <- list(
  name = "npi",
  time_begin = 30,
  time_end = 60,
  contact_reduction = matrix(0.1, 3)
)

as.intervention(npi)
}
