% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{as.vaccination}
\alias{as.vaccination}
\title{Convert a list to a vaccination object}
\usage{
as.vaccination(x)
}
\arguments{
\item{x}{A list, or an object that inherits from a list.}
}
\value{
A \link{vaccination} class object.
}
\description{
Convert a list to a vaccination object
}
\examples{
# prepare a list
vax <- list(
  name = "vax_regime",
  time_begin = matrix(1),
  time_end = matrix(100),
  nu = matrix(0.001)
)

as.vaccination(vax)
}
