% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check_helpers.R
\name{assert_infection}
\alias{assert_infection}
\title{Assert properties of an \code{infection} object}
\usage{
assert_infection(x, extra_parameters, extra_parameters_limits)
}
\arguments{
\item{x}{An \link{infection} object.}

\item{extra_parameters}{A character vector giving the names of any extra
parameters included in \code{x}. \code{x} may already be expected to have the members
\code{r0} and \code{infectious_period}.}

\item{extra_parameters_limits}{An optional named list of the numeric limits
of the acceptable limits of the parameters specifies in \code{extra_parameters}.
Each list element must be a named, two element vector of 'lower' and 'upper'
limits. Not all extra parameters need to have their limits passed explicitly.
Parameters whose limits are not explicitly specified will have their values
checked against a lower bound of 0, with the upper limit expected to be
finite.}
}
\value{
Silently returns the \code{infection} object \code{x}. Primarily called for its
side effects of throwing errors when \code{x} does not meet certain requirements.
}
\description{
Assert that objects of the \code{infection} class have the parameters expected by
an epidemic model. See \code{\link[=infection]{infection()}} and \code{\link[=epidemic]{epidemic()}}, as well as model
details to check the infection parameters required by each model. This
function is for internal use in argument checking functions.
}
\keyword{internal}
