% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check_helpers.R
\name{assert_population}
\alias{assert_population}
\title{Assert properties of a \code{population} object}
\usage{
assert_population(x, demography_groups = NULL, compartments)
}
\arguments{
\item{x}{A \link{population} object.}

\item{compartments}{A character vector giving the names of model compartments
whose length is taken as the reference for the number of columns in the
\code{initial_conditions} matrix in \code{x}.}
}
\value{
Silently returns the \code{population} object \code{x}.
Primarily called for its side effects of throwing errors when \code{x} does not
meet certain requirements.
}
\description{
Assert that objects of the \code{population} class have the parameters expected by
an epidemic model. See \code{\link[=population]{population()}} and specific epidemic functions
to check the population parameters required by each model. This
function is for internal use in argument checking functions.
}
\keyword{internal}
