% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check_helpers.R
\name{assert_vaccination}
\alias{assert_vaccination}
\title{Assert properties of a \code{vaccination} object}
\usage{
assert_vaccination(x, doses, population)
}
\arguments{
\item{x}{A \link{vaccination} object.}

\item{doses}{The number of doses expected in the vaccination object.}

\item{population}{An optional argument which is a \link{population} object.
When present, this is used to check whether the vaccination object \code{x} has
corresponding values for each demographic group in \code{population}.}
}
\value{
Silently returns the \code{vaccination} object \code{x}.
Primarily called for its side effects of throwing errors when \code{x} does not
meet certain requirements.
}
\description{
Assert that objects of the \code{vaccination} class have the parameters expected
by an epidemic model. See \code{\link[=vaccination]{vaccination()}} and \code{\link[=epidemic]{epidemic()}}, as well as model
details to check the vaccination properties required by each model. This
function is for internal use in argument checking functions.
}
\keyword{internal}
