% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{c.intervention}
\alias{c.intervention}
\title{Concatenate intervention doses into a multi-dose intervention}
\usage{
\method{c}{intervention}(x, ...)
}
\arguments{
\item{x}{An \code{intervention} object.}

\item{...}{intervention objects to combine with \code{x} to create a multi-dose
\code{intervention} object.}
}
\value{
An \code{intervention} object with as many doses as the overall number of
doses specified in \code{x} and in the objects passed to \code{...}.
}
\description{
Concatenate intervention doses into a multi-dose intervention
}
\examples{
# create first dose regime
npi_1 <- intervention(
  time_begin = 30,
  time_end = 60,
  contact_reduction = matrix(0.1)
)

# second dose regime
npi_2 <- intervention(
  time_begin = 45,
  time_end = 75,
  contact_reduction = matrix(0.1)
)

c(npi_1, npi_2)
}
