% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{c.vaccination}
\alias{c.vaccination}
\title{Concatenate vaccination doses into a multi-dose vaccination}
\usage{
\method{c}{vaccination}(x, ...)
}
\arguments{
\item{x}{A \code{vaccination} object.}

\item{...}{Vaccination objects to combine with \code{x} to create a multi-dose
\code{vaccination} object.}
}
\value{
A \code{vaccination} object with as many doses as the overall number of
doses specified in \code{x} and in the objects passed to \code{...}.
}
\description{
Concatenate vaccination doses into a multi-dose vaccination
}
\examples{
# create first dose regime
vax_1 <- vaccination(
  name = "vax_regime",
  time_begin = matrix(1),
  time_end = matrix(100),
  nu = matrix(0.001)
)

# second dose regime
vax_2 <- vaccination(
  name = "vax_regime",
  time_begin = matrix(101),
  time_end = matrix(200),
  nu = matrix(0.001)
)

c(vax_1, vax_2)
}
