% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_default.R
\name{check_args_default}
\alias{check_args_default}
\title{Check arguments to default epidemic function}
\usage{
check_args_default(mod_args)
}
\description{
Check arguments to default epidemic function
}
\keyword{internal}
