% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_default.R
\name{check_prepare_default_args}
\alias{check_prepare_default_args}
\alias{.check_args_epidemic_default}
\alias{.prepare_args_epidemic_default}
\title{Check arguments to default epidemic function}
\usage{
.check_args_epidemic_default(mod_args)

.prepare_args_epidemic_default(mod_args)
}
\value{
\code{.check_args_epidemic_default()} invisibly returns the model arguments passed
in \code{mod_args}. If the model arguments did not previously contain elements
named \code{intervention} and \code{vaccination}, these are added as dummy objects of
the corresponding classes.

\code{.prepare_args_epidemic_default()} returns a list of model arguments suitable
for \code{\link[=.epidemic_default_cpp]{.epidemic_default_cpp()}}. This is a named list consisting of:
\itemize{
\item \code{initial_state}: the initial conditions modified to represent absolute
rather than proportional values;
\item \code{beta}, \code{alpha}, \code{gamma}: three numbers representing the transmission rate
of the infection, the rate of transition from exposed to infectious, and the
recovery rate, respectively;
\item \code{contact_matrix}, a numeric matrix for the population contact matrix
scaled by the largest real eigenvalue and by the size of each groups;
\item \code{npi_time_begin}, \code{npi_time_end}: two vectors for the start and end times
of any interventions applied;
\item \code{npi_cr}: a matrix for the age- and intervention-specific effect on social
contacts;
\item \code{vax_time_begin},\code{vax_time_end}, \code{vax_nu}: three numeric matrices for the
age- and dose-specific start times, end times, and rates of any vaccination
doses implemented;
\item \code{time_end}, \code{increment}: two numbers for the time at which to end the
simulation, and the value by which the simulation time
is incremented.
}
}
\description{
Check and prepare the four main arguments to
\code{\link[=epidemic_default_cpp]{epidemic_default_cpp()}} for use with \code{\link[=.epidemic_default_cpp]{.epidemic_default_cpp()}}.

\code{.check_args_epidemic_default()} adds an empty \verb{<intervention>} and
\verb{<vaccination>} object if these are missing from the model arguments.

\code{.prepare_args_epidemics_default()} prepares arguments for
\code{\link[=.epidemic_default_cpp]{.epidemic_default_cpp()}}, which is the C++ function that solves the default
ODE system using a Boost \emph{odeint} solver.
\code{.prepare_args_epidemics_default()} converts the arguments collected in
\code{mod_args} into simpler structures such as lists and numeric or integer
vectors that can be interpreted as C++ types such as \code{Rcpp::List},
\code{Rcpp::NumericVector}, or \code{Eigen::MatrixXd}.
}
\keyword{internal}
