% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_ebola.R
\name{check_prepare_ebola_args}
\alias{check_prepare_ebola_args}
\alias{.check_args_epidemic_ebola}
\alias{.prepare_args_epidemic_ebola}
\title{Check arguments to epidemic ebola model from Getz and Dougherty}
\usage{
.check_args_epidemic_ebola(mod_args)

.prepare_args_epidemic_ebola(mod_args)
}
\value{
\code{.check_args_epidemic_ebola()} invisibly returns the model arguments passed
in \code{mod_args}.

\code{.prepare_args_epidemic_ebola()} returns a list of model arguments suitable
for \code{\link[=.model_ebola_cpp]{.model_ebola_cpp()}}. This is a named list consisting of:
\itemize{
\item \code{initial_state}: the initial conditions modified to represent absolute
rather than proportional values;
\item \code{population_size}: the population size;
\item \code{beta}: the transmission rate of the infection;
\item \code{shape_E}, \code{rate_E}: the shape and rate parameters of an Erlang
distribution that represents passage times through the 'exposed' compartment;
\item \code{shape_I}, \code{rate_I}: the shape and rate parameters of an Erlang
distribution that represents passage times through the 'infectious'
compartment;
\item \code{time_end}: the time at which the simulation ends.
}
}
\description{
Check and prepare the four main arguments to
\code{\link[=model_ebola_cpp]{model_ebola_cpp()}} for use with \code{\link[=.model_ebola_cpp]{.model_ebola_cpp()}}.
}
\keyword{internal}
