% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_population.R
\name{combine_contact}
\alias{combine_contact}
\title{Create the connectivity matrix between all groups of \code{populations}}
\usage{
combine_contact(populations, connectivity_matrix)
}
\arguments{
\item{populations}{A list of \verb{<population>} objects}

\item{connectivity_matrix}{A numeric matrix for the proportion of contacts
between the elements of the populations}
}
\value{
A numeric matrix combining the contact matrices in \verb{<population>} and
the connectivity matrix.
}
\description{
Create the connectivity matrix between all groups of \code{populations}
}
\keyword{internal}
