% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_ebola.R
\name{compute_erlang_discrete_prob}
\alias{compute_erlang_discrete_prob}
\title{Discrete probabilities for an Erlang distribution}
\usage{
compute_erlang_discrete_prob(k, gamma)
}
\arguments{
\item{k}{The shape parameter of the Erlang distribution.}

\item{gamma}{The rate parameter of the Erlang distribution.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ A vector containing all p_i values, for i = 1 : n.
}\if{html}{\out{</div>}}
}
\description{
Discrete probabilities for an Erlang distribution
}
