% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernels.R
\name{Contact_Kernel}
\alias{Contact_Kernel}
\title{Contact_Kernel Given a matrix containing information on whether individuals in an
               epidemic network are able to make infectious contact with eachother
               and returns a function which can be given infectious process parameters
               to return infectious pressure between individuals}
\usage{
Contact_Kernel(contact_matrix, SYM = TRUE)
}
\arguments{
\item{contact_matrix}{A Matrix whose ij-th entry specifies whether an infected indiviual i
can make infectious contact with susceptible individual j}

\item{SYM}{A logical arguement. When set to TRUE, it will assume that contact
is mutual, giving a symmetrical matrix. If FALSE, it mutual infectious
contact is not assumed.}
}
\description{
Contact_Kernel Given a matrix containing information on whether individuals in an
               epidemic network are able to make infectious contact with eachother
               and returns a function which can be given infectious process parameters
               to return infectious pressure between individuals
}
