% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cross_checking_inputs}
\alias{cross_checking_inputs}
\alias{.cross_check_intervention}
\alias{.cross_check_vaccination}
\alias{.cross_check_timedep}
\alias{.cross_check_popchange}
\title{Cross-check model elements}
\usage{
.cross_check_intervention(x, population, allowed_targets)

.cross_check_vaccination(x, population, doses)

.cross_check_timedep(x, allowed_targets)

.cross_check_popchange(x, population)
}
\arguments{
\item{x}{Model input to be checked. The expected value of \code{x} depends on the
function:
\itemize{
\item \code{.cross_check_intervention()}: A named list of \verb{<intervention>} objects;
\item \code{.cross_check_vaccination()}: A \verb{<vaccination>} object;
\item \code{.cross_check_timedep()}: A named list of functions with two arguments,
\code{time} and \code{x}, typically returning \code{x} as a function of \code{time};
\item \code{.cross_check_popchange()}: A named list with two elements, \code{time} and
\code{values}, describing the times and values by which the number of susceptibles
changes in an epidemic model.
}}

\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{allowed_targets}{The model components, or infection parameters, that
the model input \code{x} affects.}

\item{doses}{The expected number of vaccination doses.}
}
\value{
\itemize{
\item \code{.cross_check_intervention()} returns a named list with at least the
elements "contacts" describing a \verb{<contacts_intervention>} on \code{population}
(if this is among the allowed targets), and a \verb{<rate_intervention>} on the
transmission rate parameter. If these are present in \code{x}, they are
returned as is, or substituted if missing. Any other interventions
are also returned. If \code{x} is \code{NULL}, dummy contact and rate interventions
are returned in a list.
\item \code{.cross_check_vaccination()} returns \code{x} after checking that it is suitable
for \code{population}, or a dummy vaccination regime with \code{doses} number of doses
for each age group.
\item \code{.cross_check_timedep()} returns \code{x} if \code{x} is not \code{NULL}, otherwise
returns a dummy function operating on the transmission rate parameter by
default; see \code{\link[=.no_time_dependence]{.no_time_dependence()}};
\item \code{.cross_check_popchange()} returns \code{x} after checks against \code{population} if
\code{x} is not \code{NULL}, otherwise returns a dummy list with no population change;
see \code{\link[=.no_population_change]{.no_population_change()}}.
}
}
\description{
Check model elements for compatibility with the population in an epidemic
model, returning compatible dummy values when model elements are not applied,
and erroring appropriately when model elements are not compatible with the
population characteristics.
}
\keyword{internal}
