% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{cumulative_intervention}
\alias{cumulative_intervention}
\alias{intervention_on_cm}
\title{Calculate the Cumulative Effect of Interventions on Social Contacts}
\usage{
cumulative_intervention(t, time_begin, time_end, cr)

intervention_on_cm(t, cm, time_begin, time_end, cr)
}
\arguments{
\item{t}{The current time.}

\item{time_begin}{A numeric vector of the start times of all interventions
being modelled.}

\item{time_end}{A numeric vector of the end times of all interventions being
modelled. Must be the same length as \code{time_begin}.}

\item{cr}{A numeric matrix where rows give the effect of interventions on
each demographic group, and the columns give the proportion reduction in
contacts. When two interventions overlap, the proportions are \emph{added}, for
a maximum possible value of 1.0 (i.e., no contacts).}

\item{cm}{A numeric matrix of social contacts between demographic groups.}
}
\value{
\code{cumulative_intervention()} returns a numeric vector of the proportion
reduction in contacts for each demographic group.

\code{intervention_on_cm()} returns the contact matrix \code{cm} scaled by the
cumulative effect of any active interventions.
}
\description{
Calculate the Cumulative Effect of Interventions on Social Contacts

Scale a Contact Matrix by all Active Interventions on Social Contacts
}
\keyword{internal}
