% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{cumulative_rate_intervention}
\alias{cumulative_rate_intervention}
\title{Calculate the Cumulative Effect of Interventions on Rate Parameters}
\usage{
cumulative_rate_intervention(t, time_begin, time_end, reduction)
}
\arguments{
\item{t}{The current time.}

\item{time_begin}{A numeric vector of the start times of all interventions
being modelled.}

\item{time_end}{A numeric vector of the end times of all interventions being
modelled. Must be the same length as \code{time_begin}.}

\item{reduction}{A numeric vector where each element gives the effect of the
corresponding intervention on model rate parameters. When two interventions
overlap, the proportions are \emph{added}, for a maximum possible value of 1.0
(i.e., rate set to zero).}
}
\value{
\code{cumulative_rate_intervention()} returns a number of the proportion
reduction in a model rate parameter.

\code{intervention_on_cm()} returns the contact matrix \code{cm} scaled by the
cumulative effect of any active interventions.
}
\description{
Calculate the Cumulative Effect of Interventions on Rate Parameters
}
\keyword{internal}
