% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_default.R
\name{.check_args_epidemic_default}
\alias{.check_args_epidemic_default}
\title{Check arguments to default epidemic function}
\usage{
.check_args_epidemic_default(mod_args)
}
\value{
Invisibly returns the model arguments passed in \code{mod_args}. If the
model arguments did not previously contain elements named \code{intervention} and
\code{vaccination}, these are added as dummy objects of the corresponding classes.
}
\description{
Checks the four main arguments to \code{\link[=epidemic_default_cpp]{epidemic_default_cpp()}}.
Adds an empty \verb{<intervention>} and \verb{<vaccination>} object if these are
missing from the model arguments, and this is done for compatibility with
the model code.
}
\keyword{internal}
