% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_ebola.R
\name{.check_args_epidemic_ebola}
\alias{.check_args_epidemic_ebola}
\title{Check arguments to epidemic ebola model from Getz and Dougherty}
\usage{
.check_args_epidemic_ebola(mod_args)
}
\description{
Check arguments to epidemic ebola model from Getz and Dougherty
}
\keyword{internal}
