% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_vacamole.R
\name{.check_args_epidemic_vacamole}
\alias{.check_args_epidemic_vacamole}
\title{Check arguments to epidemic function for the Vacamole model}
\usage{
.check_args_epidemic_vacamole(mod_args)
}
\description{
Check arguments to epidemic function for the Vacamole model
}
\keyword{internal}
