% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{.cross_check_intervention}
\alias{.cross_check_intervention}
\title{Cross-check an intervention list}
\usage{
.cross_check_intervention(x, population, allowed_targets)
}
\arguments{
\item{x}{A named list of \verb{<intervention>} objects.}

\item{population}{A \verb{<population>} to which any \verb{<contact_intervention>} in
\code{x} applies, and with which it must be compatible.}

\item{allowed_targets}{A character vector of targets allowed for the
\verb{<intervention>}s in \code{x}. May contain "contacts", and the names of any
infection parameters.}
}
\value{
A named list with at least the elements "contacts" describing a
\verb{<contacts_intervention>} on \code{population}, and a \verb{<rate_intervention>} on
the transmissibility parameter. If these are present in \code{x}, they are
returned as is, or substituted if missing such that the function output is
suitable for processing for a C++ model function. Any other interventions
are also returned. If \code{x} is \code{NULL}, dummy
contact and rate interventions are returned in a list.
}
\description{
Cross-check an intervention list
}
\keyword{internal}
