% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{.cross_check_timedep}
\alias{.cross_check_timedep}
\title{Cross-check time-dependence}
\usage{
.cross_check_timedep(x, allowed_targets)
}
\arguments{
\item{x}{A named list of functions that describe how an infection parameter
changes with time.}

\item{allowed_targets}{A character vector of targets for \code{x}; these are
typically infection parameters.}
}
\value{
If \code{x} is not \code{NULL}, returns \code{x}; otherwise a dummy function is
returned.
}
\description{
Cross-check time-dependence
}
\keyword{internal}
