% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{.cross_check_vaccination}
\alias{.cross_check_vaccination}
\title{Cross-check a \verb{<vaccination>}}
\usage{
.cross_check_vaccination(x, population, doses)
}
\arguments{
\item{x}{A \verb{<vaccination>}}

\item{population}{A \verb{<population>}}

\item{doses}{A single number for the expected number of doses.}
}
\value{
Returns \code{x} after checking that it is suitable for \code{population}, or
a dummy vaccination regime with \code{doses} number of doses for each age group.
}
\description{
Cross-check a \verb{<vaccination>}
}
\keyword{internal}
