% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.epidemic_default_cpp}
\alias{.epidemic_default_cpp}
\title{Run an age-structured SEIR-V epidemic model}
\usage{
.epidemic_default_cpp(
  population,
  beta,
  alpha,
  gamma,
  intervention,
  vaccination,
  time_end = 200,
  increment = 0.1
)
}
\arguments{
\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{beta}{The transmission rate \eqn{\beta}.}

\item{alpha}{The rate of transition from exposed to infectious \eqn{\alpha}.}

\item{gamma}{The recovery rate \eqn{\gamma}.}

\item{intervention}{A non-pharamaceutical intervention applied during the
course of the epidemic, with a start and end time, and age-specific effect
on contacts. See \code{\link[=intervention]{intervention()}}.}

\item{vaccination}{A vaccination regime followed during the
course of the epidemic, with a start and end time, and age-specific effect
on the transition of individuals from susceptible to vaccinated.
See \code{\link[=vaccination]{vaccination()}}.}

\item{time_end}{The maximum time, defaults to 200.0.}

\item{increment}{The increment time, defaults to 0.1.}
}
\value{
A two element list, where the first element is a list of matrices
whose elements correspond to the numbers of individuals in each compartment
as specified in the initial conditions matrix (see \code{\link[=population]{population()}}).
The second list element is a vector of timesteps.
}
\description{
A compartmental model with an optional non-pharmaceutical
intervention and an optional vaccination regime. Allows heterogeneity in
social contact patterns, and variable sizes of demographic groups.
Also allows for group-specific initial proportions in each model
compartment, as well as group-specific vaccination start dates and
vaccination rates, and also group-specific effects of implementing a
non-pharmaceutical intervention.
The model only allows for single, population-wide rates of
transition between the 'susceptible' and 'exposed' compartments, between the
'exposed' and 'infectious' compartments, and in the recovery rate.
}
\keyword{internal}
