% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.epidemic_vacamole_cpp}
\alias{.epidemic_vacamole_cpp}
\title{Run the RIVM Vacamole model}
\usage{
.epidemic_vacamole_cpp(
  population,
  beta,
  beta_v,
  alpha,
  omega,
  omega_v,
  eta,
  eta_v,
  gamma,
  intervention,
  vaccination,
  time_end,
  increment
)
}
\arguments{
\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{beta}{The transmission rate \eqn{\beta} at which unvaccinated and
partially vaccinated individuals are infected by the disease.}

\item{beta_v}{The transmission rate \eqn{\beta_V} at which individuals who
have received two vaccine doses are infected by the disease.}

\item{alpha}{The rate of transition from exposed to infectious \eqn{\alpha}.
This is common to fully susceptible, partially vaccinated, and fully
vaccinated individuals (where fully vaccinated represents two doses).}

\item{omega}{The mortality rate of fully susceptible and partially
vaccinated and unprotected individuals.}

\item{omega_v}{The mortality rate of individuals who are protected by
vaccination.}

\item{eta}{The hospitalisation rate of fully susceptible and partially
vaccinated and unprotected individuals.}

\item{eta_v}{The hospitalisation rate of individuals who are protected by
vaccination.}

\item{gamma}{The recovery rate \eqn{\gamma}.}

\item{intervention}{A non-pharmaceutical intervention applied during the
course of the epidemic, with a start and end time, and age-specific effect
on contacts. See \code{\link[=intervention]{intervention()}}.}

\item{vaccination}{A vaccination regime followed during the
course of the epidemic, with a group- and dose-specific start and end time,
and age-specific rates of delivery of first and second doses.
See \code{\link[=vaccination]{vaccination()}}.}

\item{time_end}{The maximum time. See \code{\link[=epidemic]{epidemic()}} for default value.}

\item{increment}{The increment time. See \code{\link[=epidemic]{epidemic()}} for default value.}
}
\value{
A two element list, where the first element is a list of matrices
whose elements correspond to the numbers of individuals in each compartment
as specified in the initial conditions matrix (see \code{\link[=population]{population()}}).
The second list element is a vector of timesteps.
}
\description{
Vacamole is a deterministic, compartmental epidemic model built
by Kylie Ainslie and others at RIVM, the Dutch Public Health Institute for
the Covid-19 pandemic, with a focus on scenario modelling for
hospitalisation and vaccination.
Model code: https://github.com/kylieainslie/vacamole
Manuscript describing the model and its application:
https://doi.org/10.2807/1560-7917.ES.2022.27.44.2101090
}
\details{
The original model has 8 conceptual compartments - four
epidemiological compartments (SEIR), three hospitalisation compartments
(H, ICU, ICU2H), and death - see the manuscript in Eurosurveillance.
Only infected individuals can enter the hospitalisation or death
compartments.
Vacamole was implemented as a stand-alone R package, and some versions have
been used to generate scenarios for the ECDC Covid-19 Scenario Hub.

Individuals from the susceptible compartment may be vaccinated partially
or fully (assuming a two dose regimen), with only the second dose reducing
their probability of being infected, and of being hospitalised or dying.
}
\keyword{internal}
