% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_class.R
\name{.get_model_parameters}
\alias{.get_model_parameters}
\title{Get model arguments as a list, including default arguments}
\usage{
.get_model_parameters(f)
}
\arguments{
\item{f}{A function or a function name as a string.}
}
\description{
Internal function intended to be called inside a call to \verb{model_*_*()} to
capture values passed to function arguments, including default values.
}
\keyword{internal}
