% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.model_default_cpp}
\alias{.model_default_cpp}
\title{Run an age-structured SEIR-V epidemic ODE model using a Boost solver}
\usage{
.model_default_cpp(
  initial_state,
  transmission_rate,
  infectiousness_rate,
  recovery_rate,
  contact_matrix,
  npi_time_begin,
  npi_time_end,
  npi_cr,
  vax_time_begin,
  vax_time_end,
  vax_nu,
  rate_interventions,
  time_dependence,
  time_end = 100,
  increment = 1
)
}
\arguments{
\item{initial_state}{A matrix for the initial state of the compartments.}

\item{transmission_rate}{The transmission rate \eqn{\beta} at which
unvaccinated and partially vaccinated individuals are infected by the
disease.}

\item{infectiousness_rate}{The rate of transition from exposed to infectious
\eqn{\alpha}.}

\item{recovery_rate}{The recovery rate \eqn{\gamma}.}

\item{contact_matrix}{The population contact matrix.}

\item{npi_time_begin}{The start time of any non-pharmaceutical interventions
.}

\item{npi_time_end}{The end time of any non-pharmaceutical interventions.}

\item{npi_cr}{The reduction in contacts from any non-pharmaceutical
interventions.}

\item{vax_time_begin}{The start time of any vaccination campaigns.}

\item{vax_time_end}{The end time of any vaccination campaigns.}

\item{vax_nu}{The vaccination rate of any vaccination campaigns.}

\item{rate_interventions}{A named list of \verb{<rate_intervention>} objects.}

\item{time_dependence}{A named list of functions for parameter time
dependence.}

\item{time_end}{The end time of the simulation.}

\item{increment}{The time increment of the simulation.}
}
\value{
A two element list, where the first element is a list of matrices
whose elements correspond to the numbers of individuals in each compartment
as specified in the initial conditions matrix (see \code{\link[=population]{population()}}).
The second list element is a vector of timesteps.
}
\description{
A compartmental model with an optional non-pharmaceutical
intervention and an optional vaccination regime.

This function is intended to only be called internally from
\code{\link[=model_default_cpp]{model_default_cpp()}}.
}
\keyword{internal}
