% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.model_diphtheria_cpp}
\alias{.model_diphtheria_cpp}
\title{Run an SEIHR ODE model for diphtheria using a Boost solver}
\usage{
.model_diphtheria_cpp(
  initial_state,
  transmission_rate,
  infectiousness_rate,
  recovery_rate,
  reporting_rate,
  prop_hosp,
  hosp_entry_rate,
  hosp_exit_rate,
  rate_interventions,
  time_dependence,
  pop_change_times,
  pop_change_values,
  time_end = 100,
  increment = 1
)
}
\arguments{
\item{initial_state}{A matrix for the initial state of the compartments.}

\item{transmission_rate}{The transmission rate \eqn{\beta}.}

\item{infectiousness_rate}{The rate of transition from exposed to infectious
\eqn{\alpha}.}

\item{recovery_rate}{The recovery rate \eqn{\gamma}.}

\item{reporting_rate}{The recovery rate \eqn{\r}.}

\item{prop_hosp}{The proportion of individuals hospitalised \eqn{\eta}.}

\item{hosp_entry_rate}{The rate at which individuals are hospitalised,
represented as 1 / time to hospitalisation \eqn{\tau_1}.}

\item{hosp_exit_rate}{The rate at which individuals are discharged from
hospital, represented as 1 / time to discharge \eqn{\tau_2}.}

\item{rate_interventions}{A named list of \verb{<rate_intervention>} objects.}

\item{time_dependence}{A named list of functions for parameter time
dependence.}

\item{pop_change_times}{A numeric vector of times and which the population
of susceptibles changes.}

\item{pop_change_values}{An Rcpp List of numeric vectors giving the value of
changes to each demographic group at each change in population.}

\item{time_end}{The end time of the simulation.}

\item{increment}{The time increment of the simulation.}
}
\value{
A two element list, where the first element is a list of matrices
whose elements correspond to the numbers of individuals in each compartment
as specified in the initial conditions matrix.
The second list element is a vector of timesteps.
}
\description{
A compartmental model for diphtheria with parameters to help
account for case reporting rate, delays in seeking hospitalisation, and the
time spent in the hospitalised compartment.

This function is intended to only be called internally from
\code{\link[=model_diphtheria_cpp]{model_diphtheria_cpp()}}.
}
\keyword{internal}
