% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.model_vacamole_cpp}
\alias{.model_vacamole_cpp}
\title{Run the RIVM Vacamole model}
\usage{
.model_vacamole_cpp(
  initial_state,
  transmission_rate,
  transmission_rate_vax,
  infectiousness_rate,
  mortality_rate,
  mortality_rate_vax,
  hospitalisation_rate,
  hospitalisation_rate_vax,
  recovery_rate,
  contact_matrix,
  npi_time_begin,
  npi_time_end,
  npi_cr,
  vax_time_begin,
  vax_time_end,
  vax_nu,
  rate_interventions,
  time_dependence,
  time_end = 100,
  increment = 1
)
}
\arguments{
\item{initial_state}{A matrix for the initial state of the compartments.}

\item{transmission_rate}{The transmission rate \eqn{\beta} at which
unvaccinated and partially vaccinated individuals are infected by the
disease.}

\item{transmission_rate_vax}{The transmission rate \eqn{\beta_V} at which
individuals who have received two vaccine doses are infected by the disease.}

\item{infectiousness_rate}{The rate of transition from exposed to infectious
\eqn{\alpha}.
This is common to fully susceptible, partially vaccinated, and fully
vaccinated individuals (where fully vaccinated represents two doses).}

\item{mortality_rate}{The mortality rate of fully susceptible and partially
vaccinated and unprotected individuals.}

\item{mortality_rate_vax}{The mortality rate of individuals who are
protected by vaccination.}

\item{hospitalisation_rate}{The hospitalisation rate of fully susceptible
and partially vaccinated and unprotected individuals.}

\item{hospitalisation_rate_vax}{The hospitalisation rate of individuals who
are protected by vaccination.}

\item{recovery_rate}{The recovery rate \eqn{\gamma}.}

\item{contact_matrix}{The population contact matrix.}

\item{npi_time_begin}{The start time of any non-pharmaceutical interventions
.}

\item{npi_time_end}{The end time of any non-pharmaceutical interventions.}

\item{npi_cr}{The reduction in contacts from any non-pharmaceutical
interventions.}

\item{vax_time_begin}{The start time of any vaccination campaigns.}

\item{vax_time_end}{The end time of any vaccination campaigns.}

\item{vax_nu}{The vaccination rate of any vaccination campaigns.}

\item{rate_interventions}{A named list of \verb{<rate_intervention>} objects.}

\item{time_dependence}{A named list of functions for parameter time
dependence.}

\item{time_end}{The end time of the simulation.}

\item{increment}{The time increment of the simulation.}
}
\value{
A two element list, where the first element is a list of matrices
whose elements correspond to the numbers of individuals in each compartment
as specified in the initial conditions matrix (see \code{\link[=population]{population()}}).
The second list element is a vector of timesteps.
}
\description{
Vacamole is a deterministic, compartmental epidemic model built
by Kylie Ainslie and others at RIVM, the Dutch Public Health Institute for
the Covid-19 pandemic, with a focus on scenario modelling for
hospitalisation and vaccination.
Model code: https://github.com/kylieainslie/vacamole
Manuscript describing the model and its application:
https://doi.org/10.2807/1560-7917.ES.2022.27.44.2101090
This function is intended to only be called internally from
\code{\link[=model_vacamole_cpp]{model_vacamole_cpp()}}.
}
\keyword{internal}
