% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_elements.R
\name{.no_population_change}
\alias{.no_population_change}
\title{Specify no population change during an epidemic}
\usage{
.no_population_change(population)
}
\arguments{
\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}
}
\description{
Specify no population change during an epidemic
}
\keyword{internal}
