% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_default.R
\name{.ode_epidemic_default}
\alias{.ode_epidemic_default}
\title{Ordinary Differential Equations for the Default Model}
\usage{
.ode_epidemic_default(t, y, params)
}
\arguments{
\item{t}{A single number of the timestep at which to integrate.}

\item{y}{The conditions of the epidemiological compartments.}

\item{params}{The parameters, passed as a named list.}
}
\value{
A list with a vector with as many elements as the number of
demographic groups times the number of epidemiological compartments. Each
value gives the change in the number of individuals in that compartment.
}
\description{
Provides the ODEs for the default SEIR-V model in a format that
is suitable for passing to \code{\link[deSolve:lsoda]{deSolve::lsoda()}}.
See \code{\link[=epidemic_default_r]{epidemic_default_r()}} for a list of required parameters.
}
\keyword{internal}
