% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_ebola.R
\name{.prepare_args_epidemic_ebola}
\alias{.prepare_args_epidemic_ebola}
\title{Prepare arguments to epidemic ebola model from Getz and Dougherty}
\usage{
.prepare_args_epidemic_ebola(mod_args)
}
\description{
Prepare arguments to epidemic ebola model from Getz and Dougherty
}
\keyword{internal}
