% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_vacamole.R
\name{.prepare_args_epidemic_vacamole}
\alias{.prepare_args_epidemic_vacamole}
\title{Prepare arguments for the Vacamole epidemic function}
\usage{
.prepare_args_epidemic_vacamole(mod_args)
}
\description{
Prepare arguments for the Vacamole epidemic function
}
\keyword{internal}
