% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_stochastic_seir.R
\name{.prepare_interventions}
\alias{.prepare_interventions}
\title{.prepare_interventions}
\usage{
.prepare_interventions(contact_matrix, rates, time_end, interventions)
}
\arguments{
\item{contact_matrix}{The base contact matrix}

\item{rates}{The base transition rates}

\item{time_end}{The length of the simulation}

\item{interventions}{The interverntions applied}
}
\value{
A list of the contact matrix and transition rates applicable for
each step of the simulation
}
\description{
Converts interventions in to time dependent changes in parameters
and transmission rates
}
\keyword{internal}
