% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{.prepare_population}
\alias{.prepare_population}
\title{Prepare a \verb{<population>} for an epidemic model}
\usage{
.prepare_population(x)
}
\arguments{
\item{x}{A \verb{<population>}.}
}
\value{
A named list of "contact_matrix", the population social contacts
matrix scaled by the largest real eigenvalue and the demography vector, and
"initial_state", the proportional initial state multiplied by the demography
vector.
}
\description{
Prepare a \verb{<population>} for an epidemic model
}
\keyword{internal}
