% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.prob_discrete_erlang}
\alias{.prob_discrete_erlang}
\title{Compute the discrete probability of the truncated Erlang distribution}
\usage{
.prob_discrete_erlang(shape = 5L, rate = 1)
}
\arguments{
\item{shape}{A single integer-like number for the shape of the Erlang
distribution.}

\item{rate}{A single number for the rate of the Erlang distribution.}
}
\value{
A vector of variable length giving the probability of each integer
value for a cumulative probability of 0.99.
}
\description{
A helper function that gives the probability of discrete values
from an Erlang distribution with a given shape and rate. The number of
values returned correspond to the number of discrete values over which the
cumulative probability reaches 0.99.
}
