% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{.recycle_vectors}
\alias{.recycle_vectors}
\title{Recycle vectors with Tidyverse recycling rules}
\usage{
.recycle_vectors(x)
}
\arguments{
\item{x}{A list of vectors to be recycled. Two cases are envisaged: all
elements of \code{x} are expected to be of the same length, or, all but one
element of \code{x} are scalars.}
}
\value{
A list of vectors with all elements recycled to the same length.
}
\description{
Recycle vectors with Tidyverse recycling rules
}
\details{
Note that this function will return vectors of unequal lengths if \code{x} does
not comply with length rules. This compliance is not enforced as this
function is only expected to be used internally.
}
\keyword{internal}
