% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.sir_stochastic}
\alias{.sir_stochastic}
\title{Stochastic, continuous time SIR epidemic simulation}
\usage{
.sir_stochastic(parameters)
}
\arguments{
\item{parameters}{A named list of parameters to the simulation.}
}
\description{
An initial implementation of an SIR epidemic simulation.
}
