% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{.transpose_base}
\alias{.transpose_base}
\title{Transpose a list}
\usage{
.transpose_base(x)
}
\arguments{
\item{x}{A list of vectors or list-like elements, which have the same length.}
}
\value{
A list of the same length as each element of \code{x}, where each element
is of the length of \code{x}. When \code{x} is named, each element of the returned list
takes the name of the original element of \code{x}.
}
\description{
This is intended to be a base R implementation of the \code{transpose()}
function from the \code{{purrr}} package.
}
\author{
Tim Taylor
}
\keyword{internal}
