% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_demic.R
\name{epi_list_to_df}
\alias{epi_list_to_df}
\title{Tidy SIR model output}
\usage{
epi_list_to_df(x)
}
\arguments{
\item{x}{The output of an SIR model. Must have the list elements, as vectors,
\code{time}, \code{S}, \code{I}, and \code{R}.}
}
\value{
A data frame in tidy (long) format with the columns \code{time}, \code{S}, \code{I}
and \code{R} for the proportions of individuals infected at each time point.
}
\description{
Tidy SIR model output
}
\keyword{internal}
