% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_class.R
\name{epidemic}
\alias{epidemic}
\alias{is_epidemic}
\title{Check whether an object is an \verb{<epidemic>}}
\usage{
is_epidemic(x)
}
\arguments{
\item{x}{An object to be checked as belonging to the \verb{<epidemic>} class.}
}
\description{
Check whether an object is an \verb{<epidemic>}
}
