% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_class.R
\name{epidemic_class}
\alias{epidemic_class}
\alias{epidemic}
\title{Create an \verb{<epidemic>} object}
\usage{
epidemic(model_fn, data, parameters)
}
\arguments{
\item{model_fn}{A string giving the name of the model function.}

\item{data}{A \verb{<data.frame>} of the model output. Must have four columns
named "time", "demography_group", "compartment", and value.}

\item{parameters}{A named list of model parameters used in a model run. These
may of atomic or more complex types, such as \verb{<intervention>}, or NULL for
unused, optional arguments.}
}
\description{
A simple class to hold the output of epidemic models along with the model
parameters and any composable elements, such as interventions, applied.
}
\examples{
# Run a model to get an <epidemic> class output
uk_population <- population(
  name = "UK population",
  contact_matrix = matrix(1),
  demography_vector = 67e6,
  initial_conditions = matrix(
    c(0.9999, 0.0001, 0, 0, 0),
    nrow = 1, ncol = 5L
  )
)

# run epidemic simulation with no vaccination or intervention
output <- model_default_cpp(
  population = uk_population
)

output

# access parameters
get_parameter(output, "transmissibility")
}
