% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_epidemic_default.R
\name{epidemic_default}
\alias{epidemic_default}
\title{SIR-V epidemic model.}
\usage{
epidemic_default(t, init, params)
}
\arguments{
\item{t}{The times.}

\item{init}{The initial conditions.}

\item{params}{The parameters.}
}
\value{
A list with numeric elements, giving the status of each compartment.
}
\description{
SIR-V epidemic model.
}
\keyword{internal}
