% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{epidemic_default_cpp}
\alias{epidemic_default_cpp}
\title{An SIR model}
\usage{
epidemic_default_cpp(init, beta, gamma)
}
\arguments{
\item{init}{The initial conditions.}

\item{beta}{The transmission rate \eqn{\beta}.}

\item{gamma}{The recovery rate \eqn{\gamma}.}
}
\description{
An SIR model
}
