% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_ebola.R
\name{epidemic_ebola}
\alias{epidemic_ebola}
\alias{epidemic_ebola_r}
\alias{epidemic_ebola_cpp}
\title{Model a stochastic epidemic with Erlang passage times}
\usage{
epidemic_ebola_r(initial_state, parameters, time_end = 100)

epidemic_ebola_cpp(population, infection, time_end = 100)
}
\arguments{
\item{initial_state}{A vector that contains 4 numbers corresponding to the
initial values of the 4 classes: S, E, I, and R.}

\item{parameters}{A vector that contains 5 numbers corresponding to the
following parameters:
\enumerate{
\item \code{shape_E}, a single integer value for the shape parameter of the Erlang
distribution from which passage times are drawn for the 'exposed'
compartment.
\item \code{rate_E}, a single integer value for the rate parameter of the Erlang
distribution from which passage times are drawn for the 'exposed'
compartment.
\item \code{shape_I}, a single integer value for the shape parameter of the Erlang
distribution from which passage times are drawn for the 'infectious'
compartment.
\item \code{rate_I}, a single integer value for the rate parameter of the Erlang
distribution from which passage times are drawn for the 'infectious'
compartment.
\item \code{beta}, a single number for the transmission rate of the infection.
}}

\item{time_end}{The maximum number of timesteps over which to run the model.
Taken as days, with a default value of 100 days.}

\item{population}{An object of the \verb{<population>} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{infection}{An \verb{<infection>} object created using \code{\link[=infection]{infection()}}. Must
have the basic reproductive number \eqn{R_0} of the infection, and the
infectious period.
These are used to calculate the transmission rate \eqn{\beta}, the rate
at which individuals move from the 'exposed' to the 'infectious' compartment.
This differs from how \eqn{\beta} is passed directly as a parameter in
\code{epidemic_ebola_r()}.

The \verb{<infection>} object must hold the parameters of the Erlang distributions
of passage times through the "exposed" and "infectious" compartments:
\code{shape_E}, \code{rate_E}, \code{shape_I}, and \code{rate_I}.}
}
\value{
For \code{epidemic_ebola_r()}, a \verb{<data.frame>} containing the numbers of
individuals in each compartment, "susceptible", "exposed", "infectious", and
"recovered", over time (from 1 to \code{time_end}).

For \code{epidemic_ebola_cpp()}, a \verb{<data.table>} in long format with the columns
"time", "compartment", "age_group", and "value", that gives the number of
individuals in each model compartment over time (from 0 to \code{time_end}).
}
\description{
Simulate an epidemic using a stochastic an SEIR compartmental
model with Erlang passage times based on Getz and Dougherty (2017) in
J. Biological Dynamics, and developed to model the West African
Ebola virus disease outbreak of 2014. See \strong{Details} for more information.
}
\details{
The R code for this model is taken from code by Hạ Minh Lâm and initially
made available on \emph{Epirecipes} (https://github.com/epirecipes/epicookbook)
under the MIT licence. The model is based on Getz and Dougherty (2017); see
\strong{References}.

\code{epidemic_ebola_cpp()} is a wrapper function for
\code{\link[=.epidemic_ebola_cpp]{.epidemic_ebola_cpp()}}, a C++ function that is exposed to R via \emph{Rcpp}.
The internal C++ function \code{\link[=.epidemic_ebola_cpp]{.epidemic_ebola_cpp()}} accepts arguments that
are created by processing the \code{population}, \code{infection}, \code{intervention} and
\code{vaccination} arguments to the wrapper function into simpler forms. This
processing is performed internally.
}
\references{
Getz, W. M., & Dougherty, E. R. (2018). Discrete stochastic analogs of Erlang
epidemic models. Journal of Biological Dynamics, 12(1), 16–38.
\doi{10.1080/17513758.2017.1401677}
}
