% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_ebola.R
\name{epidemic_ebola_cpp}
\alias{epidemic_ebola_cpp}
\title{Model a stochastic epidemic with Erlang passage times}
\usage{
epidemic_ebola_cpp(population, infection, time_end = 100)
}
\arguments{
\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{infection}{An \code{infection} object created using \code{\link[=infection]{infection()}}. Must
have the basic reproductive number \eqn{R_0} of the infection, and the
infectious period.
These are used to calculate the transmission rate \eqn{\beta}, the rate
at which individuals move from the 'exposed' to the 'infectious' compartment.

This model also requires the following extra arguments in the \verb{<infection>}
object:
\enumerate{
\item \code{shape_E}, a single integer value for the shape parameter of the Erlang
distribution from which passage times are drawn for the 'exposed'
compartment.
\item \code{rate_E}, a single integer value for the rate parameter of the Erlang
distribution from which passage times are drawn for the 'exposed'
compartment.
\item \code{shape_I}, a single integer value for the shape parameter of the Erlang
distribution from which passage times are drawn for the 'infectious'
compartment.
\item \code{rate_I}, a single integer value for the rate parameter of the Erlang
distribution from which passage times are drawn for the 'infectious'
compartment.
}}

\item{time_end}{The maximum number of timesteps over which to run the model.
Taken as days, with a default value of 100 days.}
}
\value{
A \code{data.table} with the columns "time", "compartment", "age_group",
"value". The compartments correspond to the compartments of the model
chosen with \code{model}.
The current default model has the compartments "susceptible", "exposed",
"infectious", and "recovered".
}
\description{
Simulate an epidemic using a stochastic an SEIR compartmental
model with Erlang passage times based on Getz and Dougherty (2017) in
J. Biological Dynamics, and developed to model the West African
Ebola virus disease outbreak of 2014. See \strong{Details} for more information.
}
\details{
This is a wrapper function for \code{\link[=.epidemic_ebola_cpp]{.epidemic_ebola_cpp()}}, a C++
function that uses Rcpp for its calculations.

The internal C++ function \code{\link[=.epidemic_default_cpp]{.epidemic_default_cpp()}} accepts arguments that
are created by processing the \code{population}, \code{infection}, \code{intervention} and
\code{vaccination} arguments to the wrapper function into simpler forms.
}
