% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_ebola.R
\name{epidemic_ebola_r}
\alias{epidemic_ebola_r}
\title{Model a stochastic epidemic with Erlang passage times}
\usage{
epidemic_ebola_r(initial_state, parameters, max_time)
}
\arguments{
\item{initial_state}{A vector that contains 4 numbers corresponding to the
initial values of the 4 classes: S, E, I, and R.}

\item{parameters}{A vector that contains 5 numbers corresponding to the
following parameters: the shape and the rate parameters
of the Erlang distribution that will be used to
calculate the transition rates between the E components
(i.e. \code{k[E]} and \code{gamma[E]}), the shape and the rate parameters
of the Erlang distribution that will be used to
calculate the transition rates between the I components
(i.e. \code{k[I]} and \code{gamma[I]}), and the base transmission rate
(i.e. beta).}

\item{max_time}{The length of the simulation.}
}
\value{
A data frame containing the values of S, E, I, and R over time
(from 1 to max_time).
}
\description{
This is an R-only implementation of \code{\link[=epidemic_ebola_cpp]{epidemic_ebola_cpp()}}.
}
