% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIS_fsMCMC.R
\name{Epidemic_fsMCMC}
\alias{Epidemic_fsMCMC}
\title{Endemic (SIS) fsMCMC}
\usage{
Epidemic_fsMCMC(
  N,
  a,
  x,
  beta0,
  gamma0,
  kernel = NULL,
  no_draws,
  s,
  T_obs,
  k,
  lambda,
  V,
  no_its,
  burn_in = 0,
  lag_max = NA,
  thinning_factor = 1
)
}
\arguments{
\item{N}{Total size of closed population}

\item{x}{panel data observed. Follows a random sample of n individuals and observes them at k timepoints}

\item{beta0}{beta starting value (Infection Rate Parameter)}

\item{gamma0}{gamma starting value (Removal rate parameter)}

\item{no_draws}{How many rexp(1) and runif(1) draws to make for use in the Gillespie algorithm.}

\item{T_obs}{the period for which the epidemic is observed.}

\item{k}{how many equally spaced observations take place}

\item{lambda}{RWM proposal parameter}

\item{V}{RWM proposal covariance matrix}

\item{no_its}{The number of MCMC iterations}

\item{burn_in}{How many of the MCMC iterations are thrown away as burn in (Convergence to Stationary Distn)}

\item{lag_max}{When plotting the estimated ACF of samples, what will be the maximum lag estimated/plotted}

\item{thinning_factor}{Create Storage Matrix
Proposal Acceptance Counter
Propose new beta and gamma using Multiplicative RW propsal
Draw New Random Variables
Store State
Calculating Summary Statistics for samples}

\item{initial_infective}{Number of individuals who are initially infected in the population}
}
\description{
Epidemic_fsMCMC MCMC algorithm to make inference on panel data of an epidemic through forward simulation
}
