% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_Calculations.R
\name{epidemic_loglikelihood}
\alias{epidemic_loglikelihood}
\title{Computes the full epidemic likelihood}
\usage{
epidemic_loglikelihood(par_inf, par_rem, t_inf, t_rem, kernel)
}
\description{
Computes the full epidemic likelihood
}
