% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vacamole.R
\name{epidemic_vacamole}
\alias{epidemic_vacamole}
\alias{model_vacamole_cpp}
\alias{model_vacamole_r}
\title{Model leaky, two-dose vaccination in an epidemic using Vacamole}
\usage{
model_vacamole_cpp(
  population,
  infection,
  intervention = NULL,
  vaccination,
  time_dependence = NULL,
  time_end = 100,
  increment = 1
)

model_vacamole_r(
  population,
  infection,
  intervention = NULL,
  vaccination,
  time_dependence = NULL,
  time_end = 100,
  increment = 1
)
}
\arguments{
\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{infection}{An \code{infection} object created using \code{\link[=infection]{infection()}}. Must
have the basic reproductive number \eqn{R_0} of the infection, the
infectious period, and the pre-infectious period.

These are used to calculate the transmission rate \eqn{\beta}, the rate
at which individuals move from the 'exposed' to the 'infectious' compartment,
\eqn{\alpha}, and the recovery rate \eqn{\gamma}.

This model also requires the \verb{<infection>} object to have:
\enumerate{
\item A single parameter \code{eta} for the hospitalisation rate of infectious
individuals,
\item A single parameter \code{omega} for the mortality rate of infectious or
hospitalised individuals,
\item A single parameter \code{susc_reduction_vax}, with a value between 0.0 and 1.0
giving the reduction in susceptibility to infection of individuals who have
received two doses of the vaccine,
\item A single parameter \code{hosp_reduction_vax}, with a value between 0.0 and 1.0
giving the reduction in hospitalisation rates of infectious individuals who
have received two doses of the vaccine,
\item A single parameter \code{mort_reduction_vax}, with a value between 0.0 and 1.0
giving the reduction in mortality of infectious and hospitalised individuals
who have received two doses of the vaccine.
}}

\item{intervention}{An \verb{<intervention>} object representing an optional
non-pharmaceutical intervention applied to the population during the
epidemic. See \code{\link[=intervention]{intervention()}} for details on constructing interventions with
age-specific effects on social contacts, as well as for guidance on how to
concatenate multiple overlapping interventions into a single \verb{<intervention>}
object.}

\item{vaccination}{A \verb{<vaccination>} object representing an optional
vaccination regime with two doses followed during the course of the
epidemic, with a start and end time, and age-specific vaccination rates for
each dose. See \code{\link[=vaccination]{vaccination()}}.}

\item{time_dependence}{A named list where each name
is a model parameter (see \code{infection}), and each element is a function with
the first two arguments being the current simulation \code{time}, and \code{x}, a value
that is dependent on \code{time} (\code{x} represents a model parameter).
See \strong{Details} for more information, as well as the vignette on time-
dependence \code{vignette("time_dependence", package = "epidemics")}.}

\item{time_end}{The maximum number of timesteps over which to run the model.
Taken as days, with a default value of 200 days.}

\item{increment}{The size of the time increment. Taken as days, with a
default value of 1 day.}
}
\value{
A \code{data.table} with the columns "time", "compartment", "age_group",
"value". The compartments correspond to the compartments of the model
chosen with \code{model}.
The current default model has the compartments "susceptible",
"vaccinated_one_dose", "vaccinated_two_dose", "exposed",
"infectious", "infectious_vaccinated", "hospitalised",
"hospitalised_vaccinated", "recovered",  and "dead".
}
\description{
Simulate an epidemic using the \emph{Vacamole} model for Covid-19
developed at RIVM, the National Institute for Public Health and the
Environment in the Netherlands.
This model is aimed at estimating the impact of 'leaky' vaccination on an
epidemic. See \strong{Details} for more information.
}
\details{
This model allows for:
\enumerate{
\item A 'hospitalised' compartment along with hospitalisation rates;
\item Two doses of vaccination, with 'leaky' protection, i.e., vaccination does
not prevent infection completely but allows for a reduction in the infection
rate, as well as reduced rates of moving into states considered more serious,
such as 'hospitalised' or 'dead'.
}

\code{model_vacamole_cpp()} is a wrapper function for
\code{\link[=.model_vacamole_cpp]{.model_vacamole_cpp()}}, a C++ function that uses Boost \emph{odeint} solvers
to implement the RIVM Vacamole model.

\code{model_vacamole_r()} is a wrapper around the internal function
\code{.ode_model_vacamole()}, which is passed to \code{\link[deSolve:lsoda]{deSolve::lsoda()}}.

Both models return equivalent results, but the C++ implementation is faster.

\code{.model_vacamole_cpp()} and \code{.ode_model_vacamole()} both accept
arguments that are created by processing the \code{population}, \code{infection},
\code{intervention} and \code{vaccination} arguments to the wrapper function into
simpler forms.

Model rates or parameters can be made time-dependent by passing a function
which modifies the parameter based on the current ODE simulation time.
For example, a function that modifies the transmission rate \code{beta} could be
passed as \code{time_dependence = list(beta = function(time, x) x + sinpi(time))}.
This functionality may be used to model events that are expected to have some
effect on model parameters, such as seasonality or annual schedules such as
holidays.
}
\references{
Ainslie, K. E. C., Backer, J. A., Boer, P. T. de, Hoek, A. J. van,
Klinkenberg, D., Altes, H. K., Leung, K. Y., Melker, H. de, Miura, F., &
Wallinga, J. (2022). A scenario modelling analysis to anticipate the impact
of COVID-19 vaccination in adolescents and children on disease outcomes in
the Netherlands, summer 2021. Eurosurveillance, 27(44), 2101090.
\doi{10.2807/1560-7917.ES.2022.27.44.2101090}
}
