% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_vacamole.R
\name{epidemic_vacamole_cpp}
\alias{epidemic_vacamole_cpp}
\title{Model leaky, two-dose vaccination in an epidemic using Vacamole}
\usage{
epidemic_vacamole_cpp(
  population,
  infection,
  intervention = NULL,
  vaccination,
  time_end = 100,
  increment = 1
)
}
\arguments{
\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{infection}{An \code{infection} object created using \code{\link[=infection]{infection()}}. Must
have the basic reproductive number \eqn{R_0} of the infection, the
infectious period, and the pre-infectious period.

These are used to calculate the transmission rate \eqn{\beta}, the rate
at which individuals move from the 'exposed' to the 'infectious' compartment,
\eqn{\alpha}, and the recovery rate \eqn{\gamma}.

This model also requires the \verb{<infection>} object to have:
\enumerate{
\item A single parameter \code{eta} for the hospitalisation rate of infectious
individuals,
\item A single parameter \code{omega} for the mortality rate of infectious or
hospitalised individuals,
\item A single parameter \code{susc_reduction_vax}, with a value between 0.0 and 1.0
giving the reduction in susceptibility to infection of individuals who have
received two doses of the vaccine,
\item A single parameter \code{hosp_reduction_vax}, with a value between 0.0 and 1.0
giving the reduction in hospitalisation rates of infectious individuals who
have received two doses of the vaccine,
\item A single parameter \code{mort_reduction_vax}, with a value between 0.0 and 1.0
giving the reduction in mortality of infectious and hospitalised individuals
who have received two doses of the vaccine.
}}

\item{intervention}{An \verb{<intervention>} object representing an optional
non-pharmaceutical intervention applied to the population during the
epidemic. See \code{\link[=intervention]{intervention()}} for details on constructing interventions with
age-specific effects on social contacts, as well as for guidance on how to
concatenate multiple overlapping interventions into a single \verb{<intervention>}
object.}

\item{vaccination}{A \verb{<vaccination>} object representing an optional
vaccination regime with two doses followed during the course of the
epidemic, with a start and end time, and age-specific vaccination rates for
each dose.
See \code{\link[=vaccination]{vaccination()}}.}

\item{time_end}{The maximum number of timesteps over which to run the model.
Taken as days, with a default value of 200 days.}

\item{increment}{The size of the time increment. Taken as days, with a
default value of 1 day.}
}
\value{
A \code{data.table} with the columns "time", "compartment", "age_group",
"value". The compartments correspond to the compartments of the model
chosen with \code{model}.
The current default model has the compartments "susceptible",
"vaccinated_one_dose", "vaccinated_two_dose", "exposed",
"infectious", "infectious_vaccinated", "hospitalised",
"hospitalised_vaccinated", "recovered",  and "dead".
}
\description{
Simulate an epidemic using the \emph{Vacamole} model for Covid-19
developed at RIVM, the National Institute for Public Health and the
Environment in the Netherlands.
This model is aimed at estimating the impact of 'leaky' vaccination on an
epidemic. See \strong{Details} for more information.
}
\details{
This is a wrapper function for \code{\link[=.epidemic_vacamole_cpp]{.epidemic_vacamole_cpp()}}, a C++
function that uses Boost \emph{odeint} solvers for an SEIHR-V model.

This model allows for:
\enumerate{
\item A 'hospitalised' compartment along with hospitalisation rates;
\item Two doses of vaccination, with 'leaky' protection, i.e., vaccination does
not prevent infection completely but allows for a reduction in the infection
rate, as well as reduced rates of moving into states considered more serious,
such as 'hospitalised' or 'dead'.
}

The internal C++ function \code{\link[=.epidemic_vacamole_cpp]{.epidemic_vacamole_cpp()}} accepts arguments that
are created by processing the \code{population}, \code{infection}, \code{intervention} and
\code{vaccination} arguments to the wrapper function into simpler forms.
}
