% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsMCMC_SIS.R
\name{fsMCMC_SIS}
\alias{fsMCMC_SIS}
\title{Create Storage Matrix
Proposal Acceptance Counter
Propose new beta and gamma using Multiplicative RW propsal
Store State
Calculating Summary Statistics for samples}
\usage{
fsMCMC_SIS(
  obsTransData,
  obsTimes,
  N,
  beta0,
  gamma0,
  lambda,
  s,
  noIts,
  burnIn = 0,
  lagMax = NA,
  thinningFactor = 1
)
}
\description{
Create Storage Matrix
Proposal Acceptance Counter
Propose new beta and gamma using Multiplicative RW propsal
Store State
Calculating Summary Statistics for samples
}
