% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Epidemic_Simulation.R
\name{General_Epidemic_Process}
\alias{General_Epidemic_Process}
\title{The parameters and the rate structure together will determine the rates of the
process for each individual}
\usage{
General_Epidemic_Process(
  parameters,
  no_states,
  state_names,
  rate_structure,
  initial_population
)
}
\description{
States should be ordered in such a way which follows the flow of the epidemic.
}
\details{
Therefore the rate structure should be ordered in the same way

SIR Example:

State 1 ---> S, State 2 ---> I, State 3 ---> R

1 ---> 2,  Beta Parameters

2 ---> 3,  Gamma Parameters
}
