% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_helpers.R
\name{get_model_names}
\alias{get_model_names}
\title{Get names of available models}
\usage{
get_model_names(model_type = "epidemic")
}
\arguments{
\item{model_type}{The model type. The only option currently available is
"epidemic". Future development may see the addition of more model types,
including endemic disease compartmental models, and branching process
outbreak models.}
}
\description{
Get names of available models
}
