% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_parameter}
\alias{get_parameter}
\title{Get parameters from \{epidemics\} classes}
\usage{
get_parameter(x, parameter)
}
\arguments{
\item{x}{An object of one of the classes provided by \emph{epidemics}:
\verb{<population>}, \verb{<infection>}, \verb{<intervention>}, or \verb{<vaccination>}.}

\item{parameter}{A string for the parameter to access from \code{x}.}
}
\value{
An object of the class of \code{parameter} from \code{x}.
}
\description{
Get parameters from \{epidemics\} classes
}
