% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection.R
\name{get_transmission_rate}
\alias{get_transmission_rate}
\title{Calculate the transmission rate of an \verb{<infection>}}
\usage{
get_transmission_rate(infection)
}
\arguments{
\item{infection}{An \verb{<infection>} object, with the parameters \code{r0} for the
basic reproductive number, and \code{infectious_period}, for the mean infectious
period.}
}
\value{
A single number representing the transmission rate of the infection,
\eqn{\beta}.
}
\description{
Calculate the transmission rate of an \verb{<infection>}
}
\examples{
# create an infection
ebola <- infection(r0 = 1.3, infectious_period = 5, preinfectious_period = 5)
get_transmission_rate(infection = ebola)
}
