% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Homo_LowPop_MedR0}
\alias{Homo_LowPop_MedR0}
\title{Simulated Epidemic Dataset}
\format{R list of 7:
\describe{
  \item{N}{Size of Closed Population, in People}
  \item{final_size}{Number of individuals infected during course of the
                    epidemic, excluding the initial infected}
  \item{t_rem}{Removal Times of those removed during the course of the
               the epidemic, in a arbitrary unit of time}
  \item{t_inf}{Infection Times of those infected during the course of
               the epidemic}
  \item{beta}{Infectious Process Parameter}
  \item{gamma}{Infectious Period Parameter}
  \item{R0}{Basic Reproductive Rate}
  }}
\source{
{Epidemic_Simulation Function}
}
\usage{
Homo_LowPop_MedR0
}
\description{
Dataset containing the infection and removal times of a simulated homogeneous
epidemic with closed population N = 100, infectious process parameter beta = 0.0024
and infectious period parameter gamma = 0.15. This epidemic has a small population
with a medium reproductive rate compared to the other simulated datasets in this
package.
}
\keyword{datasets}
