% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_gillespies.R
\name{homogeneousPanelDataSIR_Gillespie}
\alias{homogeneousPanelDataSIR_Gillespie}
\title{}
\usage{
homogeneousPanelDataSIR_Gillespie(initialState, beta, gamma, obsTimes)
}
\description{
currentTime = 0
B = kernel(beta)
#' Possible states and their respective
#' logical representations.
susceptible = c(TRUE, FALSE, FALSE)
infected    = c(FALSE, TRUE, FALSE)
removed     = c(FALSE, FALSE, TRUE)
}
\details{

}
