% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_gillespies.R
\name{homogeneousPanelDataSIR_GillespieEU}
\alias{homogeneousPanelDataSIR_GillespieEU}
\title{Possible states and their respective
logical representations.}
\usage{
homogeneousPanelDataSIR_GillespieEU(initialState, beta, gamma, obsTimes, E, U)
}
\description{
Summarise in matrix
Initial state of each individual
Initial state of epidemic at population level
Possible Events
Summarise in a matrix
event table
Waiting Time
Panel Data
Convert individualStates into panelData
Update Population State
Update Individual State
Filling in Observation times not passed
Convert individualStates into panelData
}
