% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIS_gillespies.R
\name{homogeneousPanelDataSIS_Gillespie}
\alias{homogeneousPanelDataSIS_Gillespie}
\title{==== SIS Gillespie with Panel Data Output ====
Possible states and their respective
logical representations.
Summarise in matrix
Initial state of each individual
Initial state of epidemic at population level
Possible Events
Summarise in a matrix
event table
Waiting Time
Panel Data
Convert individualStates into panelData
This which does take up significant time
#' Which event happens?
#' To who?
Update Population State
Update Individual State
Filling in Observation times not passed
Convert individualStates into panelData}
\usage{
homogeneousPanelDataSIS_Gillespie(initialState, beta, gamma, obsTimes)
}
\description{
==== SIS Gillespie with Panel Data Output ====
Possible states and their respective
logical representations.
Summarise in matrix
Initial state of each individual
Initial state of epidemic at population level
Possible Events
Summarise in a matrix
event table
Waiting Time
Panel Data
Convert individualStates into panelData
This which does take up significant time
#' Which event happens?
#' To who?
Update Population State
Update Individual State
Filling in Observation times not passed
Convert individualStates into panelData
}
