% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIS_gillespies.R
\name{homogeneousPanelDataSIS_GillespieEU}
\alias{homogeneousPanelDataSIS_GillespieEU}
\title{==== Non-Centered SIS Gillespie with Panel Data Output ====
Possible states and their respective
logical representations.
Summarise in matrix
Initial state of each individual
Initial state of epidemic at population level
Possible Events
Summarise in a matrix
event table
Waiting Time
Panel Data
Convert individualStates into panelData
To who?
Update Population State
Update Individual State
Filling in Observation times not passed
Convert individualStates into panelData}
\usage{
homogeneousPanelDataSIS_GillespieEU(initialState, beta, gamma, obsTimes, E, U)
}
\description{
==== Non-Centered SIS Gillespie with Panel Data Output ====
Possible states and their respective
logical representations.
Summarise in matrix
Initial state of each individual
Initial state of epidemic at population level
Possible Events
Summarise in a matrix
event table
Waiting Time
Panel Data
Convert individualStates into panelData
To who?
Update Population State
Update Individual State
Filling in Observation times not passed
Convert individualStates into panelData
}
