% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_gillespies.R
\name{homogeneousSIR_Gillespie}
\alias{homogeneousSIR_Gillespie}
\title{SIR Gillespie *NEW*
Split into seperate cases
State is (X, Y, Z)
Input
Initial State (X_0, Y_0, Z_0)
Initialise state
Possible events are infection and removal
Construct event matrix
Create event table
Calculate rate of each possible event
Total rate will dictate the waiting time
distribution.
Which event occured?
event = sample(c(1, 2), size = 1, prob = c(infectionRate, removalRate))
Alter state of epidemic
Update Event table}
\usage{
homogeneousSIR_Gillespie(state0, beta, gamma)
}
\description{
SIR Gillespie *NEW*
Split into seperate cases
State is (X, Y, Z)
Input
Initial State (X_0, Y_0, Z_0)
Initialise state
Possible events are infection and removal
Construct event matrix
Create event table
Calculate rate of each possible event
Total rate will dictate the waiting time
distribution.
Which event occured?
event = sample(c(1, 2), size = 1, prob = c(infectionRate, removalRate))
Alter state of epidemic
Update Event table
}
