% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIS_gillespies.R
\name{individualHomogeneousSIS_Gillespie}
\alias{individualHomogeneousSIS_Gillespie}
\title{Possible states and their respective
logical representations.
Summarise in matrix
Initial state of each individual
Initial state of epidemic at population level
Possible Events
Summarise in a matrix
event table
Waiting Time
#' Which event happens?
#' To who?
Update Population State
Update Individual State}
\usage{
individualHomogeneousSIS_Gillespie(initialState, beta, gamma, obsTimes)
}
\description{
Possible states and their respective
logical representations.
Summarise in matrix
Initial state of each individual
Initial state of epidemic at population level
Possible Events
Summarise in a matrix
event table
Waiting Time
#' Which event happens?
#' To who?
Update Population State
Update Individual State
}
