% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centered_MCMC.R
\name{inf_time_proposal}
\alias{inf_time_proposal}
\title{PhD
A Gibbs Sampler for a partially observed SIR epidemic.}
\usage{
inf_time_proposal(t_inf, t_rem, no_proposals, par_rem)
}
\description{
A Centered Parameterisation of Bayesian Modelling of the Infection and Removal times
of an Epidemic Centered
}
\details{
==== Model ====

theta(beta, gamma) --> Infection Times (I, Unobserved) --> Removal Times (R, Observed)


Approach is based on that used by Neal & Roberts (2003).


1. Take observed data and simulate a set of valid infection times using inital values
   for gamma and beta.

   Steps for one iteration:

2. Calculate the parameters of the conditional posteriors
   to be sampled from.

3. Sample from the conditional distributions for the infection and removal rate
   in turn

4. Sample an Infection Time to change. Propose the change and Accept/Reject using
   Metropolis-Hastings Acceptance probability
}
