% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection.R
\name{infection}
\alias{infection}
\title{Prepare a new infection for an epidemic model}
\usage{
infection(name = NA_character_, r0, infectious_period, ...)
}
\arguments{
\item{name}{String for the name of the infection. The default value is \code{NA}.}

\item{r0}{The basic reproductive number of the infection, \eqn{R_0}.
Must be a number or a vector of numbers of the same length as the number of
demographic groups, depending on the model being implemented.
The 'default' model requires a single number.}

\item{infectious_period}{The mean infections period.
Must be a number or a vector of numbers of the same length as the number of
demographic groups, depending on the model being implemented.
Currently, only the option to pass a single infectious period common to all
demographic groups is available.}

\item{...}{Other arguments representing other infection parameters. This
argument is intentionally left flexible to allow for single values or vectors
of values.}
}
\value{
An object of the \code{infection} S3 class.
}
\description{
Prepare a new infection for an epidemic model
}
\examples{
infection(
  name = "pandemic influenza",
  r0 = 1.3,
  infectious_period = 5,
  preinfectious_period = 2
)
}
