% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infectious_Disease_Gillespie.R
\name{Infectious_Disease_Gillespie}
\alias{Infectious_Disease_Gillespie}
\title{Infectious_Disease_Gillespie function}
\usage{
Infectious_Disease_Gillespie(
  type = "SIR",
  N,
  a,
  beta,
  gamma,
  obs_end = Inf,
  kernel = NULL,
  U,
  E,
  output = "event",
  obs_times = NULL
)
}
\description{
Wrapper function for various Infectious Disease Process Gillespie algorithms.
}
\details{
Takes the usual arguments plus an extra argument to choose which process you
want to run i.e SIR/SIS/SEIR
}
