% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_desolve.R
\name{initial_conditions}
\alias{initial_conditions}
\title{Initial conditions for SIR models}
\usage{
initial_conditions(p_initially_infected = 0.01, p_already_recovered = 0)
}
\arguments{
\item{p_initially_infected}{The proportion initial infected.}

\item{p_already_recovered}{The proportion already recovered from infection.}
}
\value{
A named vector, ("S": susceptible, "I": infected, or "R": recovered).
}
\description{
Initial conditions for SIR models
}
\examples{
initial_conditions(p_initially_infected = 0.001)
}
