% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_Calculations.R
\name{interval_intersect}
\alias{interval_intersect}
\title{Relies on the period of time for which individual j exerts pressure on individual j
Denote this as the exposure time.}
\usage{
interval_intersect(interval_i, interval_j)
}
\value{
The length of intersection between the respective intervals in interval_i and interval_j
}
\description{
Relies on the period of time for which individual j exerts pressure on individual j
Denote this as the exposure time.
}
