% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{intervention_constructor}
\alias{intervention_constructor}
\alias{new_intervention}
\alias{new_contacts_intervention}
\alias{new_rate_intervention}
\title{Constructor for the \if{html}{\out{<intervention>}} super-class and sub-classes}
\usage{
new_intervention(
  name = NA_character_,
  time_begin,
  time_end,
  reduction,
  ...,
  class
)

new_contacts_intervention(name, time_begin, time_end, reduction)

new_rate_intervention(name, time_begin, time_end, reduction)
}
\arguments{
\item{name}{String for the name of the intervention.}

\item{time_begin}{A matrix with a single element giving the start time of the
intervention.}

\item{time_end}{A matrix with a single element giving the end time of the
intervention.}

\item{reduction}{A matrix with a single column and as many rows as there are
demographic groups to be targeted by the intervention. See \code{\link[=intervention]{intervention()}}
for details on the types of intervention that can be created and the
requirements for the type of \code{reduction}.}

\item{...}{Any other parameters to be passed to the constructor.}

\item{class}{A string giving the type of the intervention; used to generate
intervention sub-classes.}
}
\value{
\code{new_intervention()} returns an object that inherits from the
\verb{<intervention>} class.
\code{new_contacts_intervention()} returns an object that is a sub-class of
\verb{<intervention>} called \verb{<contacts_intervention>}.
\code{new_rate_intervention()} returns an object that is a sub-class of
\verb{<intervention>} called \verb{<rate_intervention>}.
}
\description{
Constructor for the \if{html}{\out{<intervention>}} super-class and sub-classes

Constructor for a new <contacts_intervention>

Constructor for a new <rate_intervention>
}
\keyword{internal}
