% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{intervention_on_params}
\alias{intervention_on_params}
\title{Apply interventions to rate parameters}
\usage{
intervention_on_params(t, interventions, parameters)
}
\arguments{
\item{t}{A single number for the simulation time.}

\item{interventions}{A named list of \code{list}-like objects that each have at
least the three members \code{"time_begin"}, \code{"time_end"}, and \code{"reduction"}.
These are used to calculate the effect on each of the named parameters in the
simulation.}

\item{parameters}{A named list of numeric parameters affected by
\code{interventions}.
This represents the infection parameters, such as the transmission rate,
\eqn{\beta}, or the recovery rate, \eqn{\gamma}.}
}
\value{
A named list of the same length as \code{parameters}, with the same names.
These parameters can then be used in a timestep of an ODE model.
}
\description{
Apply interventions to rate parameters
}
