% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection.R
\name{is_infection}
\alias{is_infection}
\title{Check whether an object is an \code{infection}}
\usage{
is_infection(object)
}
\arguments{
\item{object}{An object to be checked as being an \code{infection}.}
}
\value{
A logical for whether the object is of the \code{infection} class.
}
\description{
Check whether an object is an \code{infection}
}
\examples{
pandemic_influenza <- infection(
  name = "pandemic influenza",
  r0 = 1.3,
  infectious_period = 5,
  preinfectious_period = 2
)
is_infection(pandemic_influenza)
}
