% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{is_intervention}
\alias{is_intervention}
\title{Check whether an object is an \code{intervention}}
\usage{
is_intervention(object)
}
\arguments{
\item{object}{An object to be checked as being an \code{intervention}.}
}
\value{
A logical for whether the object is of the \code{intervention} class.
}
\description{
Check whether an object is an \code{intervention}
}
\examples{
close_schools <- intervention(
  name = "close schools",
  time_begin = 50,
  time_end = 80,
  contact_reduction = matrix(c(0.5, 0.01)) # reduces contacts differentially
)
is_intervention(close_schools)
}
