% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{is_population}
\alias{is_population}
\title{Check whether an object is a \code{population}}
\usage{
is_population(object)
}
\arguments{
\item{object}{An object to be checked as a valid population.}
}
\value{
A logical for whether the object is a \code{population}.
}
\description{
Check whether an object is a \code{population}
}
\examples{
# for the UK
new_pop <- population(
  initial_conditions = matrix(
    c(0.999, 0.001, 0, 0),
    nrow = 1, ncol = 4
  )
)
is_population(new_pop)
}
