% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{is_vaccination}
\alias{is_vaccination}
\title{Check whether an object is a \code{vaccination}}
\usage{
is_vaccination(object)
}
\arguments{
\item{object}{An object to be checked as being a \code{vaccination}.}
}
\value{
A logical for whether the object is of the \code{vaccination} class.
}
\description{
Check whether an object is a \code{vaccination}
}
\examples{
# an example for childhood vaccination only
childhood_vaccination <- vaccination(
  name = "childhood_vaccination",
  time_begin = c(0, 100), # assuming a simulation over 100 days
  time_end = c(100, 100),
  nu = c(0.0001, 0.0) # over 5s never vaccinated
)
is_vaccination(childhood_vaccination)
}
