% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_Calculations.R
\name{log_likelihood_rem}
\alias{log_likelihood_rem}
\title{Computes the Removal portion of an epidemic likelihood}
\usage{
log_likelihood_rem(par, t_inf, t_rem)
}
\arguments{
\item{par}{parameter values to be passed to gamma density function (par[1] rate parameter, par[2] shape parameter)}

\item{t_inf}{infection times of individuals involved in the epidemic}

\item{t_rem}{removal times of individuals involved in the epidemic}
}
\description{
Computes the Removal portion of an epidemic likelihood
}
