% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parameters.R
\name{make_parameters_sir}
\alias{make_parameters_sir}
\title{Make convenient parameters}
\usage{
make_parameters_sir(option = c("stochastic", "deterministic"))
}
\arguments{
\item{option}{Either "stochastic" or "deterministic".}
}
\description{
Function to make convenient parameters.
}
\examples{
make_parameters_sir(option = "stochastic")
make_parameters_sir(option = "deterministic")
}
