% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parameters.R
\name{make_parameters_sir_stochastic}
\alias{make_parameters_sir_stochastic}
\title{Make convenient parameters}
\usage{
make_parameters_sir_stochastic()
}
\description{
Function to make convenient parameters.
}
