% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Output.R
\name{MCMCplots}
\alias{MCMCplots}
\title{Functions for MCMC Output}
\usage{
MCMCplots(samples, lagMax = NA, ylab, trueValue = NULL, set.mfrow = FALSE)
}
\description{
Trace and ACF plotting function
Trace Plot
Density Plot
}
