% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_default.R
\name{model_default}
\alias{model_default}
\alias{model_default_cpp}
\alias{model_default_r}
\title{Model an SEIR-V epidemic with interventions}
\usage{
model_default_cpp(
  population,
  transmissibility = 1.3/7,
  infectiousness_rate = 1/2,
  recovery_rate = 1/7,
  intervention = NULL,
  vaccination = NULL,
  time_dependence = NULL,
  time_end = 100,
  increment = 1
)

model_default_r(
  population,
  transmissibility = 1.3/7,
  infectiousness_rate = 1/2,
  recovery_rate = 1/7,
  intervention = NULL,
  vaccination = NULL,
  time_dependence = NULL,
  time_end = 100,
  increment = 1
)
}
\arguments{
\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{transmissibility}{A single number for the rate at which individuals
move from the susceptible to the exposed compartment upon contact with an
infectious individual. Often denoted as \eqn{\beta}, with
\eqn{\beta = R_0 / \text{infectious period}}.}

\item{infectiousness_rate}{A single number for the rate at which individuals
move from the exposed to the infectious compartment. Often denoted as
\eqn{\sigma}, with \eqn{\sigma = 1.0 / \text{pre-infectious period}}.
This value does not depend upon the number of infectious individuals in the
population.}

\item{recovery_rate}{A single number for the rate at which individuals move
from the infectious to the recovered compartment. Often denoted as
\eqn{\gamma}, with \eqn{\gamma = 1.0 / \text{infectious period}}.}

\item{intervention}{An \verb{<intervention>} object representing an optional
non-pharmaceutical intervention applied to the population during the
epidemic. See \code{\link[=intervention]{intervention()}} for details on constructing interventions with
age-specific effects on social contacts, as well as for guidance on how to
concatenate multiple overlapping interventions into a single \verb{<intervention>}
object.}

\item{vaccination}{A \verb{<vaccination>} object representing an optional
vaccination regime with a single dose, followed during the course of the
epidemic, with a start and end time, and age-specific vaccination rates.
See \code{\link[=vaccination]{vaccination()}}.}

\item{time_dependence}{A named list where each name
is a model parameter, and each element is a function with
the first two arguments being the current simulation \code{time}, and \code{x}, a value
that is dependent on \code{time} (\code{x} represents a model parameter).
See \strong{Details} for more information, as well as the vignette on time-
dependence \code{vignette("time_dependence", package = "epidemics")}.}

\item{time_end}{The maximum number of timesteps over which to run the model.
Taken as days, with a default value of 200 days.}

\item{increment}{The size of the time increment. Taken as days, with a
default value of 1 day.}
}
\value{
A \code{data.table} with the columns "time", "compartment", "age_group",
"value". The compartments correspond to the compartments of the model
chosen with \code{model}.
The current default model has the compartments "susceptible", "exposed",
"infectious", "recovered", and "vaccinated".
}
\description{
Simulate an epidemic using a deterministic, compartmental
epidemic model with the compartments
"susceptible", "exposed", "infectious", "recovered", and "vaccinated".
This model can accommodate heterogeneity in social contacts among demographic
groups, as well as differences in the sizes of demographic groups.

The \code{population}, \code{transmissibility}, \code{infectiousness_rate}, and
\code{recovery_rate}
arguments are mandatory, while passing an \code{intervention} and \code{vaccination}
are optional and can be used to simulate scenarios with different epidemic
responses or different levels of the same type of response.
See \strong{Details} for more information.
}
\details{
\subsection{R and Rcpp implementations}{

\code{model_default_cpp()} is a wrapper function for \code{\link[=.model_default_cpp]{.model_default_cpp()}},
an internal C++ function that uses Boost \emph{odeint} solvers for an SEIR-V model
.
\code{\link[=.model_default_cpp]{.model_default_cpp()}} accepts arguments that
are created by processing the \code{population}, \code{infection}, \code{intervention} and
\code{vaccination} arguments to the wrapper function into simpler forms.

\code{model_default_r()} is a wrapper around the internal function
\code{.ode_model_default()}, which is passed to \code{\link[deSolve:lsoda]{deSolve::lsoda()}}.

Both models return equivalent results, but the C++ implementation is faster.
}

\subsection{Model parameters}{

This model only allows for single, population-wide rates of
transition between the 'susceptible' and 'exposed' compartments, between the
'exposed' and 'infectious' compartments, and in the recovery rate.

The default values are:
\itemize{
\item Transmissibility (\eqn{\beta}, \code{transmissibility}): 0.186, resulting from
an \eqn{R_0} = 1.3 and an infectious period of 7 days.
\item Infectiousness rate (\eqn{\sigma}, \code{infectiousness_rate}): 0.5, assuming
a pre-infectious period of 2 days.
\item Recovery rate (\eqn{\gamma}, \code{recovery_rate}): 0.143, assuming an
infectious period of 7 days.
}
}

\subsection{Implementing time-dependent parameters}{

Model rates or parameters can be made time-dependent by passing a function
which modifies the parameter based on the current ODE simulation time.
For example, a function that modifies the transmission rate
\code{transmissibility} could be passed as
\code{time_dependence = list(transmissibility = function(time, x) x * time)}.
This functionality may be used to model events that are expected to have some
effect on model parameters, such as seasonality or annual schedules such as
holidays.
}
}
