% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_default.R
\name{model_default_cpp}
\alias{model_default_cpp}
\title{Model an SEIR-V epidemic with interventions}
\usage{
model_default_cpp(
  population,
  transmission_rate = 1.3/7,
  infectiousness_rate = 1/2,
  recovery_rate = 1/7,
  intervention = NULL,
  vaccination = NULL,
  time_dependence = NULL,
  time_end = 100,
  increment = 1
)
}
\arguments{
\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{transmission_rate}{A numeric for the rate at which individuals
move from the susceptible to the exposed compartment upon contact with an
infectious individual. Often denoted as \eqn{\beta}, with
\eqn{\beta = R_0 / \text{infectious period}}. See \strong{Details} for default
values.}

\item{infectiousness_rate}{A numeric for the rate at which individuals
move from the exposed to the infectious compartment. Often denoted as
\eqn{\sigma}, with \eqn{\sigma = 1.0 / \text{pre-infectious period}}.
This value does not depend upon the number of infectious individuals in the
population. See \strong{Details} for default values.}

\item{recovery_rate}{A numeric for the rate at which individuals move
from the infectious to the recovered compartment. Often denoted as
\eqn{\gamma}, with \eqn{\gamma = 1.0 / \text{infectious period}}.
See \strong{Details} for default values.}

\item{intervention}{A named list of \verb{<intervention>}s representing optional
non-pharmaceutical or pharmaceutical interventions applied during the
epidemic. Only a single intervention on social contacts of the class
\verb{<contacts_intervention>} is allowed as the named element "contacts".
Multiple \verb{<rate_interventions>} on the model parameters are allowed; see
\strong{Details} for the model parameters for which interventions are supported.}

\item{vaccination}{A \verb{<vaccination>} object representing an optional
vaccination regime with a single dose, followed during the course of the
epidemic, with a start and end time, and age-specific vaccination rates.}

\item{time_dependence}{A named list where each name
is a model parameter, and each element is a function with
the first two arguments being the current simulation \code{time}, and \code{x}, a value
that is dependent on \code{time} (\code{x} represents a model parameter).
See \strong{Details} for more information, as well as the vignette on time-
dependence \code{vignette("time_dependence", package = "epidemics")}.}

\item{time_end}{The maximum number of timesteps over which to run the model.
Taken as days, with a default value of 100 days. May be a numeric vector.}

\item{increment}{The size of the time increment. Taken as days, with a
default value of 1 day.}
}
\value{
A \verb{<data.table>}.
If the model parameters and composable elements are all scalars, a single
\verb{<data.table>} with the columns "time", "compartment", "age_group", and
"value", giving the number of individuals per demographic group
in each compartment at each timestep in long (or "tidy") format is returned.

If the model parameters or composable elements are lists or list-like,
a nested \verb{<data.table>} is returned with a list column "data", which holds
the compartmental values described above.
Other columns hold parameters and composable elements relating to the model
run. Columns "scenario" and "param_set" identify combinations of composable
elements (population, interventions, vaccination regimes), and infection
parameters, respectively.
}
\description{
Simulate an epidemic using a deterministic, compartmental
epidemic model with the compartments
"susceptible", "exposed", "infectious", "recovered", and "vaccinated".
This model can accommodate heterogeneity in social contacts among demographic
groups, as well as differences in the sizes of demographic groups.

The \code{population}, \code{transmission_rate}, \code{infectiousness_rate}, and
\code{recovery_rate}
arguments are mandatory, while passing an \code{intervention} and \code{vaccination}
are optional and can be used to simulate scenarios with different epidemic
responses or different levels of the same type of response.
See \strong{Details} for more information.
}
\section{Details: SEIRV model suitable for directly transmitted infections}{
\subsection{Model parameters}{

This model only allows for single, population-wide rates of
transitions between compartments per model run.

However, model parameters may be passed as numeric vectors. These vectors
must follow Tidyverse recycling rules: all vectors must have the same length,
or, vectors of length 1 will be recycled to the length of any other vector.

The default values are:
\itemize{
\item Transmission rate (\eqn{\beta}, \code{transmission_rate}): 0.186, assuming an
\eqn{R_0} = 1.3 and an infectious period of 7 days.
\item Infectiousness rate (\eqn{\sigma}, \code{infectiousness_rate}): 0.5, assuming
a pre-infectious period of 2 days.
\item Recovery rate (\eqn{\gamma}, \code{recovery_rate}): 0.143, assuming an
infectious period of 7 days.
}
}
}

\examples{
# create a population
uk_population <- population(
  name = "UK population",
  contact_matrix = matrix(1),
  demography_vector = 67e6,
  initial_conditions = matrix(
    c(0.9999, 0.0001, 0, 0, 0),
    nrow = 1, ncol = 5L
  )
)

# run epidemic simulation with no vaccination or intervention
# and three discrete values of transmission rate
data <- model_default_cpp(
  population = uk_population,
  transmission_rate = c(1.3, 1.4, 1.5) / 7.0, # uncertainty in R0
)

# view some data
data

# run epidemic simulations with differences in the end time
# may be useful when considering different start dates with a fixed end point
data <- model_default_cpp(
  population = uk_population,
  time_end = c(50, 100, 150)
)

data
}
