% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diphtheria.R
\name{model_diphtheria}
\alias{model_diphtheria}
\alias{model_diphtheria_cpp}
\title{Model a diphtheria outbreak using a compartmental ODE model}
\usage{
model_diphtheria_cpp(
  population,
  transmissibility = 4/4.5,
  infectiousness_rate = 1/3,
  recovery_rate = 1/3,
  reporting_rate = 0.03,
  prop_hosp = 0.01,
  hosp_entry_rate = 0.2,
  hosp_exit_rate = 0.2,
  prop_vaccinated = 0 * get_parameter(population, "demography_vector"),
  intervention = NULL,
  time_dependence = NULL,
  population_change = NULL,
  time_end = 100,
  increment = 1
)
}
\arguments{
\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{transmissibility}{A single number for the rate at which individuals
move from the susceptible to the exposed compartment upon contact with an
infectious individual. Often denoted as \eqn{\beta}, with
\eqn{\beta = R_0 / \text{infectious period}}.}

\item{infectiousness_rate}{A single number for the rate at which individuals
move from the exposed to the infectious compartment. Often denoted as
\eqn{\sigma}, with \eqn{\sigma = 1.0 / \text{pre-infectious period}}.
This value does not depend upon the number of infectious individuals in the
population.}

\item{recovery_rate}{A single number for the recovery rate, denoted
\eqn{\gamma}.}

\item{reporting_rate}{A single number for the proportion of infectious cases
that is reported; this is a precursor to hospitalisation as only reported
cases are hospitalised.}

\item{prop_hosp}{A single number for the proportion of reported cases that is
hospitalised.}

\item{hosp_entry_rate}{A single number for the rate at which reported cases
of infectious individuals are hospitalised.
This is calculated as 1 / time to hospitalisation, denoted \eqn{\tau_1}.}

\item{hosp_exit_rate}{A single number for the rate at which individuals are
discharged from hospital to enter the 'recovered' compartment.
This is calculated as 1 / time to discharge, denoted \eqn{\tau_2}.}

\item{prop_vaccinated}{A numeric vector of the same length as the number of
demographic groups indicated the proportion of each group that is vaccinated.
These individuals are not included in the model dynamics.}

\item{intervention}{A named list of \verb{<rate_intervention>} objects
representing optional pharmaceutical or non-pharmaceutical interventions
applied to the model parameters listed above.}

\item{time_dependence}{A named list where each name
is a model parameter, and each element is a function with
the first two arguments being the current simulation \code{time}, and \code{x}, a value
that is dependent on \code{time} (\code{x} represents a model parameter).
See \strong{Details} for more information, as well as the vignette on time-
dependence \code{vignette("time_dependence", package = "epidemics")}.}

\item{population_change}{A two-element list, with elements named \code{"time"} and
\code{"values"}, giving the times of population changes, and the corresponding
changes in the population of each demographic group at those times.
\code{"time"} must be a numeric vector, while \code{"values"} must be a list of the
length of \code{"time"}, with each element a numeric vector of the same length as
the number of demographic groups in \code{population}.
default value of 1 day.}

\item{time_end}{The maximum number of timesteps over which to run the model.
Taken as days, with a default value of 100 days.}

\item{increment}{The size of the time increment. Taken as days, with a}
}
\value{
An \verb{<epidemic>} class object with the model function name,
model data as "data", model parameters and components as "parameters", and
the package version as "hash".
}
\description{
Simulate a diphtheria outbreak using a deterministic,
compartmental ordinary differential equation model with the compartments
"susceptible", "exposed", "infectious", "hospitalised", and"recovered".
The model is based on Finger et al. (2019) and is intended to be used in the
context of internally displaced people (IDP) or refugee camps.
This model accommodates age or demographic structure and allows for a
proportion of each demographic group to be vaccinated at the start of the
outbreak, and thus to not contribute to the outbreak.
The model also allows for changes to the initial population size, to model
influxes or evacuations from camps.
}
\details{
\subsection{Rcpp implementations}{

\code{model_diphtheria_cpp()} is a wrapper function for \code{\link[=.model_diphtheria_cpp]{.model_diphtheria_cpp()}},
an internal C++ function that uses Boost \emph{odeint} solvers for an SEIHR model.
}

\subsection{Model parameters}{

This model only allows for single, population-wide rates transitions between
compartments. The default values are taken from Finger et al. (2019) where
possible.
\itemize{
\item Transmissibility (\eqn{\beta}, \code{transmissibility}): 0.8888889, assuming an
\eqn{R_0} of 4.0 and a total infectious period of 4.5 days.
\item Infectiousness rate (\eqn{\sigma}, \code{infectiousness_rate}): 0.333, assuming
a pre-infectious period of 3 days.
\item Reporting rate (\eqn{r}, \code{reporting_rate}): 0.03, assuming that 3\% of
infectious cases are detected or reported.
\item Proportion hospitalised (\eqn{\eta}, \code{prop_hosp}): 0.01, assuming that 1\%
of reported cases need hospital treatment.
\item Hospital entry rate (\eqn{\tau_1}, \code{hosp_entry_rate}): 0.2, assuming that
it takes 5 days for infectious individuals to seek hospital treatment.
\item Hospital exit rate (\eqn{\tau_2}, \code{hosp_exit_rate}): 0.2, assuming that
individuals are discharged from hospital after 5 days.
\item Recovery rate (\eqn{\gamma}, \code{recovery_rate}): 0.333, assuming an
infectious period following symptoms, of 3 days.
}
}

\subsection{Modelling population changes}{

This model allows changes to the number of susceptibles in each demographic
group, to represent influxes or evacuations from the camp as would be
expected in humanitarian relief situations.
Users can specify the times and changes (to each demographic group) of
changes using the \code{population_changes} argument, to examine the effect on
outbreak dynamics.
}
}
\examples{
n_age_groups <- 3
demography_vector <- c(83000, 108200, 224600)
initial_conditions <- matrix(0, nrow = n_age_groups, ncol = 5)
# set susceptibles and infectious
initial_conditions[, 1] <- demography_vector - 1
initial_conditions[, 3] <- rep(1, n_age_groups)
camp_pop <- population(
  contact_matrix = matrix(1, nrow = n_age_groups, ncol = n_age_groups),
  demography_vector = demography_vector,
  initial_conditions = initial_conditions / demography_vector
)

model_diphtheria_cpp(camp_pop)

}
\references{
Finger, F., Funk, S., White, K., Siddiqui, M. R., Edmunds, W. J., &
Kucharski, A. J. (2019). Real-time analysis of the diphtheria outbreak in
forcibly displaced Myanmar nationals in Bangladesh. BMC Medicine, 17, 58.
\doi{10.1186/s12916-019-1288-7}.
}
