% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ebola.R
\name{model_ebola}
\alias{model_ebola}
\alias{model_ebola_r}
\title{Model an Ebola virus disease epidemic}
\usage{
model_ebola_r(
  population,
  erlang_subcompartments = 2,
  transmissibility = 1.5/12,
  infectiousness_rate = erlang_subcompartments/5,
  removal_rate = erlang_subcompartments/12,
  prop_community = 0.9,
  etu_risk = 0.7,
  funeral_risk = 0.5,
  intervention = NULL,
  time_dependence = NULL,
  time_end = 100
)
}
\arguments{
\item{population}{An object of the \verb{<population>} class, see \code{\link[=population]{population()}}.

This model only accepts a \verb{<population>} without demographic structure, that
is, the \code{demography_vector} must be a single number representing the total
size of the affected population.

The model also does not account for demographic differences in social
contacts, which means that the \code{contact_matrix} is ignored. For consistency,
the matrix must be square and have as many rows as demography groups, which
is one.}

\item{erlang_subcompartments}{The number of Erlang subcompartments assumed
for the exposed, infectious, and hospitalised compartments. Defaults to 2.}

\item{transmissibility}{A single number for the rate at which individuals
move from the susceptible to the exposed compartment upon contact with an
infectious individual. Often denoted as \eqn{\beta}, with
\eqn{\beta = R_0 / \text{infectious period}}.}

\item{infectiousness_rate}{A single number for the rate at which individuals
move from the exposed to the infectious compartment. Often denoted as
\eqn{\sigma}, with \eqn{\sigma = 1.0 / \text{pre-infectious period}}.
This value does not depend upon the number of infectious individuals in the
population.}

\item{removal_rate}{The rate at which infectious individuals transition from
the infectious or hospitalised compartments to the funeral or removed
compartments. This model does not distinguish between recoveries and
deaths. Denoted in Getz and Dougherty as \eqn{\gamma^I} (see \strong{Details}).}

\item{prop_community}{The proportion of infectious individuals who remain in
the community and are not hospitalised for treatment. Defaults to 0.5}

\item{etu_risk}{The relative risk of onward transmission of EVD from
hospitalised individuals. Must be a single value between 0.0 and 1.0, where
0.0 indicates that hospitalisation completely prevents onward transmission,
and 1.0 indicates that hospitalisation does not prevent onward transmission
at all. \code{etu_risk} is used to scale the value of transmissibility for the
transsmissibility \eqn{\beta}. Defaults to 0.2.}

\item{funeral_risk}{The relative risk of onward transmission of EVD from
funerals of individuals who died with EVD.
Must be a single value between 0.0 and 1.0, where
0.0 indicates that there is no onward transmission, and 1.0 indicates that
funeral transmisison is equivalent to transmission in the community.
\code{funeral_risk} is used to scale the value of transmissibility for the
transmissibility \eqn{\beta}. Defaults to 0.5.}

\item{intervention}{An optional \verb{<rate_intervention>} object representing
pharmaceutical or non-pharmaceutical interventions applied to the infection's
parameters, such as the transmission rate, over the epidemic.
See \code{\link[=intervention]{intervention()}} for details on constructing rate interventions.
Defaults to \code{NULL}, representing no interventions on model parameters.}

\item{time_dependence}{A named list where each name
is a model parameter (see \code{infection}), and each element is a function with
the first two arguments being the current simulation \code{time}, and \code{x}, a value
that is dependent on \code{time} (\code{x} represents a model parameter).
See \strong{Details} for more information, as well as the vignette on time-
dependence \code{vignette("time_dependence", package = "epidemics")}.}

\item{time_end}{The maximum number of timesteps over which to run the model,
in days. Defaults to 100 days.}
}
\value{
A \verb{<data.table>} in
long format with the columns "time", "compartment", "age_group", and "value",
that gives the number of individuals in each model compartment over time (
from 1 to \code{time_end}).
}
\description{
Simulate an epidemic using a discrete-time, stochastic SEIR
compartmental model with compartments based on Li et al. (2019), and with
Erlang passage times based on a model developed by Getz and Dougherty (2017),
developed to model the West African Ebola virus disease (EVD) outbreak of
2013 -- 2016.
See \strong{Details} for more information.

\code{model_ebola_cpp()} is an Rcpp implementation of this model that currently
lags behind the R implementation, and is likely to be removed.
}
\section{Details: Discrete-time ebola virus disease model}{
This model has compartments adopted from the consensus model for Ebola virus
disease presented in Li et al. (2019), and with transitions between
epidemiological compartments modelled using Erlang sub-compartments adapted
from Getz and Dougherty (2018); see \strong{References}.

The R code for this model is adapted from code by Hạ Minh Lâm and initially
made available on \emph{Epirecipes} (https://github.com/epirecipes/epicookbook)
under the MIT licence.

The shape of the Erlang distributions of passage times through the exposed
and infectious compartments (\eqn{k^E} and \eqn{k^I}) are recommended to be
set to 2 as a sensible choice, which is the default value for the
\code{erlang_sbubcompartments} argument, but can be allowed to vary
(but not independently).

The transition rates between the exposed and infectious, and infectious and
funeral compartments (and also hospitalised to removed),
\eqn{\gamma^E} and \eqn{\gamma^I} in Getz and Dougherty's notation, are
passed by the user as the \code{infectiousness_rate} and \code{removal_rate}
respectively.

Getz and Dougherty's equation (6) gives the relationship between these
parameters and the mean pre-infectious \eqn{\rho^E} and infectious
\eqn{\rho^I} periods.
\deqn{\gamma^E = \dfrac{k^E}{\rho^E} = \dfrac{2}{\rho^E} ~\text{and}~
\gamma^I = \dfrac{k^I}{\rho^I} = \dfrac{2}{\rho^I}}

In this discrete time model, \eqn{\gamma^E} and \eqn{\gamma^I} are used to
determine the passage times of newly exposed or infectious individuals
through their respective compartments (thus allowing for variation in passage
times).
\subsection{Hospitalisation, funerals, and removal}{

A proportion, \code{1.0 - prop_community}, of infectious individuals are
transferred to the hospitalised compartment in each timestep,
This compartment represents Ebola Treatment Units (ETUs), and individuals
in the hospitalised compartment are considered to be infectious but no longer
in the community.

The passage time of individuals in the hospitalised compartment is similar to
that of individuals in the infectious compartment (i.e., infectious in the
community), which means that an infectious individual with \eqn{N} timesteps
before exiting the infectious compartment will exit the hospitalised
compartment in the same time.

Hospitalised individuals can contribute to transmission of Ebola to
susceptibles depending on the value of \code{etu_risk} passed as part of the
\code{infection} argument, which scales the
baseline transmission rate \eqn{\beta} for hospitalised individuals.

We assume that deaths in hospital lead to Ebola-safe funerals, and
individuals exiting the hospitalised compartment move to the 'removed'
compartment, which holds both recoveries and deaths.

We assume that deaths outside of hospital lead to funerals that are
potentially unsafe burials, and the \code{funeral_risk} argument scales the
baseline transmission rate \eqn{\beta} for funeral transmission of Ebola to
susceptibles.

Individuals are assumed to spend only a single timestep in the funeral
transmission compartment, before they move into the 'removed' compartment.

Individuals in the 'removed' compartment do no affect model dynamics.
}

\subsection{Model parameters}{

The default values are:
\itemize{
\item Transmissibility (\eqn{\beta}, \code{transmissibility}): 0.125, resulting from
an \eqn{R_0} = 1.5 and an infectious period of 12 days.
\item Infectiousness rate (\eqn{\gamma^E}, \code{infectiousness_rate}): 0.4, assuming
a pre-infectious period of 5 days and two Erlang subcompartments.
\item Removal rate (\eqn{\gamma^I}, \code{recovery_rate}): 0.1667, assuming an
infectious period of 12 days and two Erlang subcompartments.
}
}

\subsection{Implementing vaccination}{

Vaccination cannot currently be implemented in this model as it does not have
a "vaccinated" epidemiological compartment. This prevents the use of a
\verb{<vaccination>} object.

Instead, users can use the \code{time_dependence} argument to pass a function that
modifies model parameters --- specifically, the transmission rate --- in a
way that is consistent with the effect of vaccination.
An example is shown in the vignette about this model; run this code to open
the vignette: \code{vignette("ebola_model", package = "epidemics")}
}
}

\references{
Li, S.-L., Ferrari, M. J., Bjørnstad, O. N., Runge, M. C., Fonnesbeck, C. J.,
Tildesley, M. J., Pannell, D., & Shea, K. (2019). Concurrent assessment of
epidemiological and operational uncertainties for optimal outbreak control:
Ebola as a case study. Proceedings of the Royal Society B: Biological
Sciences, 286(1905), 20190774. \doi{10.1098/rspb.2019.0774}

Getz, W. M., & Dougherty, E. R. (2018). Discrete stochastic analogs of Erlang
epidemic models. Journal of Biological Dynamics, 12(1), 16–38.
\doi{10.1080/17513758.2017.1401677}
}
