% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ebola copy.R
\name{model_ebola2}
\alias{model_ebola2}
\title{Replacement ebola model}
\usage{
model_ebola2(
  population,
  erlang_subcompartments = 2,
  transmission_rate = 1.5/12,
  infectiousness_rate = erlang_subcompartments/5,
  removal_rate = erlang_subcompartments/12,
  prop_community = 0.9,
  etu_risk = 0.7,
  funeral_risk = 0.5,
  intervention = NULL,
  time_dependence = NULL,
  time_end = 100,
  replicates = 100
)
}
\description{
Replacement ebola model
}
