% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_helpers.R
\name{model_library}
\alias{model_library}
\title{Read and display the model library}
\usage{
model_library()
}
\value{
\verb{A data.frame} representing the model library.
}
\description{
Read and display the model library
}
