% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_stochastic_seir.R
\name{model_stochastic_seir}
\alias{model_stochastic_seir}
\title{Model an stochastic SEIR epidemic with interventions}
\usage{
model_stochastic_seir(
  population,
  transmission_rate = 1.3/7,
  infectiousness_rate = 1/2,
  recovery_rate = 1/7,
  intervention = NULL,
  time_end = 100,
  n_samples = 1000
)
}
\arguments{
\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{transmission_rate}{A numeric for the rate at which individuals
move from the susceptible to the exposed compartment upon contact with an
infectious individual. Often denoted as \eqn{\beta}, with
\eqn{\beta = R_0 / \text{infectious period}}. See \strong{Details} for default
values.}

\item{infectiousness_rate}{A numeric for the rate at which individuals
move from the exposed to the infectious compartment. Often denoted as
\eqn{\sigma}, with \eqn{\sigma = 1.0 / \text{pre-infectious period}}.
This value does not depend upon the number of infectious individuals in the
population. See \strong{Details} for default values.}

\item{recovery_rate}{A numeric for the rate at which individuals move
from the infectious to the recovered compartment. Often denoted as
\eqn{\gamma}, with \eqn{\gamma = 1.0 / \text{infectious period}}.
See \strong{Details} for default values.}

\item{intervention}{A named list of \verb{<intervention>}s representing optional
non-pharmaceutical or pharmaceutical interventions applied during the
epidemic. Only a single intervention on social contacts of the class
\verb{<contacts_intervention>} is allowed as the named element "contacts".
Multiple \verb{<rate_interventions>} on the model parameters are allowed; see
\strong{Details} for the model parameters for which interventions are supported.}

\item{time_end}{The maximum number of timesteps over which to run the model.
Taken as days, with a default value of 100 days.}

\item{n_samples}{The number of stochastic replicates of the model (default = 1,000).}
}
\value{
A \verb{<data.table>}.
\verb{<data.table>} with the columns "sample", time", "compartment", "age_group",
and"value", giving the number of individuals per demographic group
in each compartment at each timestep in long (or "tidy") format is returned.
}
\description{
Simulate an epidemic using a stochastic compartmental
epidemic model with the compartments
"susceptible", "exposed", "infectious", and "recovered".
The model can accommodate heterogeneity in social contacts among demographic
groups, as well as differences in the sizes of demographic groups.
Each individual within a compartment is assumed to be independent of the
others, with inter-compartment transition times being geometrically distributed.
For exposures to infections, the compartments are assumed to be well-mixed,
thus the level of exposure felt be each individual within a compartment is
the same. Again, we assume individuals within a compartment are independent
of others in the compartment, thus the number of new infections is binomially
distributed.

The \code{population}, \code{transmission_rate}, \code{infectiousness_rate}, and
\code{recovery_rate}
arguments are mandatory, while passing an \code{intervention}
is optional and can be used to simulate scenarios with different epidemic
responses or different levels of the same type of response.
See \strong{Details} for more information.
}
\section{Details: Stochastic SEIR model suitable for directly transmitted infections}{
\subsection{Model parameters}{

This model only allows for single, population-wide rates of
transitions between compartments per model run.
Additionally, the transmission, infectiousness and recovery rates must
be scalars.

The default values are:
\itemize{
\item Transmission rate (\eqn{\beta}, \code{transmission_rate}): 0.186, assuming an
\eqn{R_0} = 1.3 and an infectious period of 7 days.
\item Infectiousness rate (\eqn{\sigma}, \code{infectiousness_rate}): 0.5, assuming
a pre-infectious period of 2 days.
\item Recovery rate (\eqn{\gamma}, \code{recovery_rate}): 0.143, assuming an
infectious period of 7 days.
}
}
}

\examples{
# create a population
uk_population <- population(
  name = "UK population",
  contact_matrix = matrix(1),
  demography_vector = 67e6,
  initial_conditions = matrix(
    c(0.9999, 0.0001, 0, 0),
    nrow = 1, ncol = 4L
  )
)

# run epidemic simulation with no vaccination or intervention
data <- model_stochastic_seir(
  population = uk_population,
  transmission_rate = 1.5 / 7.0
)

# view some data
data
}
