% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vacamole.R
\name{model_vacamole}
\alias{model_vacamole}
\alias{model_vacamole_cpp}
\alias{model_vacamole_r}
\title{Model leaky, two-dose vaccination in an epidemic using Vacamole}
\usage{
model_vacamole_cpp(
  population,
  transmissibility = 1.3/7,
  infectiousness_rate = 1/2,
  hospitalisation_rate = 1/1000,
  mortality_rate = 1/1000,
  recovery_rate = 1/7,
  susc_reduction_vax = 0.2,
  hosp_reduction_vax = 0.2,
  mort_reduction_vax = 0.2,
  intervention = NULL,
  vaccination,
  time_dependence = NULL,
  time_end = 100,
  increment = 1
)

model_vacamole_r(
  population,
  transmissibility = 1.3/7,
  infectiousness_rate = 1/2,
  hospitalisation_rate = 1/1000,
  mortality_rate = 1/1000,
  recovery_rate = 1/7,
  susc_reduction_vax = 0.2,
  hosp_reduction_vax = 0.2,
  mort_reduction_vax = 0.2,
  intervention = NULL,
  vaccination,
  time_dependence = NULL,
  time_end = 100,
  increment = 1
)
}
\arguments{
\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{transmissibility}{A single number for the rate at which individuals
move from the susceptible to the exposed compartment upon contact with an
infectious individual. Often denoted as \eqn{\beta}, with
\eqn{\beta = R_0 / \text{infectious period}}.}

\item{infectiousness_rate}{A single number for the rate at which individuals
move from the exposed to the infectious compartment. Often denoted as
\eqn{\sigma}, with \eqn{\sigma = 1.0 / \text{pre-infectious period}}.
This value does not depend upon the number of infectious individuals in the
population.}

\item{hospitalisation_rate}{A single number for the hospitalisation rate of
infectious individuals.}

\item{mortality_rate}{A single parameter for the mortality rate of
infectious or hospitalised individuals.}

\item{recovery_rate}{A single number for the rate at which individuals move
from the infectious to the recovered compartment. Often denoted as
\eqn{\gamma}, with \eqn{\gamma = 1.0 / \text{infectious period}}.}

\item{susc_reduction_vax}{A single value between 0.0 and 1.0
giving the reduction in susceptibility of infectious individuals who
have received two doses of the vaccine.}

\item{hosp_reduction_vax}{A single value between 0.0 and 1.0
giving the reduction in hospitalisation rate of infectious individuals who
have received two doses of the vaccine.}

\item{mort_reduction_vax}{A single value between 0.0 and 1.0
giving the reduction in mortality of infectious and hospitalised individuals
who have received two doses of the vaccine.}

\item{intervention}{A named list of \verb{<intervention>}s representing optional
non-pharmaceutical or pharmaceutical interventions applied during the
epidemic. Only a single intervention on social contacts of the class
\verb{<contacts_intervention>} is allowed as the named element "contacts".
Multiple \verb{<rate_interventions>} on the model parameters are allowed; see
\strong{Details} for the model parameters for which interventions are supported.}

\item{vaccination}{A \verb{<vaccination>} object representing a
vaccination regime with \strong{two doses} followed during the course of the
epidemic, with a start and end time, and age-specific vaccination rates for
each dose. See \code{\link[=vaccination]{vaccination()}}.}

\item{time_dependence}{A named list of functions that modify model parameters
as a function of model time. List element names must correspond to model
parameter names. List elements must be functions of the form
\verb{function(time, x, ...)}, where \code{time} is the simulation \code{time} and \code{x}
represents a model parameter. The order of function arguments is important.}

\item{time_end}{The maximum number of timesteps over which to run the model.
Taken as days, with a default value of 200 days.}

\item{increment}{The size of the time increment. Taken as days, with a
default value of 1 day.}
}
\value{
An \verb{<epidemic>} class object with the model function name,
model data as "data", model parameters and components as "parameters", and
the package version as "hash".
}
\description{
Simulate an epidemic using the \emph{Vacamole} model for Covid-19
developed at RIVM, the National Institute for Public Health and the
Environment in the Netherlands.
This model is aimed at estimating the impact of 'leaky' vaccination on an
epidemic. See \strong{Details} and \strong{References} for more information.
}
\details{
The Vacamole model has the compartments:
\itemize{
\item "susceptible", "exposed", "infectious", "hospitalised", "recovered", and
"dead"; as well as the special compartments
\item vaccinated with one dose and fully susceptible to infection
("vaccinated_one_dose"),
\item vaccinated with two doses and with potentially reduced susceptibility to
infection ("vaccinated_two_dose"),
\item infected and infectious after two doses of vaccination and with
potentially reduced risk of hospitalisation ("infectious_vaccinated"),
\item hospitalised after two doses of vaccination and with potentially reduced
risk of death ("hospitalised_vaccinated").
This model allows for:
}
\enumerate{
\item A 'hospitalised' compartment along with a hospitalisation rates;
\item Two doses of vaccination, with 'leaky' protection, i.e., vaccination does
not prevent infection completely but allows for a reduction in the infection
rate, as well as reduced rates of moving into states considered more serious,
such as 'hospitalised' or 'dead'.
}
\subsection{R and Rcpp implementations}{

\code{model_vacamole_cpp()} is a wrapper function for
\code{\link[=.model_vacamole_cpp]{.model_vacamole_cpp()}}, an internal C++ function that uses Boost \emph{odeint}
solvers, while \code{model_vacamole_r()} is a wrapper around \code{\link[deSolve:lsoda]{deSolve::lsoda()}}
which takes the the internal function ``.ode_model_vacamole()`.
Both models return equivalent results, but the C++ implementation is faster.
}

\subsection{Model parameters}{

This model only allows for single, population-wide rates of
transitions between compartments. The default values are:
\itemize{
\item Transmissibility (\eqn{\beta}, \code{transmissibility}): 0.186, resulting from
an \eqn{R_0} = 1.3 and an infectious period of 7 days.
\item Infectiousness rate (\eqn{\sigma}, \code{infectiousness_rate}): 0.5, assuming
a pre-infectious period of 2 days.
\item Hospitalisation rate (\eqn{\eta}, \code{hospitalistion_rate}): 1.0 / 1000,
assuming that one in every thousand infectious individuals is hospitalised.
\item Mortality rate (\eqn{\omega}, \code{mortality_rate}): 1.0 / 1000,
assuming that one in every thousand infectious and hospitalised
individuals dies.
\item Recovery rate (\eqn{\gamma}, \code{recovery_rate}): 0.143, assuming an
infectious period of 7 days.
\item Susceptibility reduction from vaccination (\code{susc_reduction_vax}): 0.2,
assuming a 20\% reduction in susceptibility for individuals who are doubly
vaccinated.
\item Hospitalisation reduction from vaccination (\code{hosp_reduction_vax}): 0.2,
assuming a 20\% reduction in hospitalisation for individuals who are doubly
vaccinated.
\item Mortality reduction from vaccination (\code{mort_reduction_vax}): 0.2,
assuming a 20\% reduction in mortality for individuals who are doubly
vaccinated.
}
}
}
\examples{
# create a population, note eleven columns for compartments
population <- population(
  contact_matrix = matrix(1),
  demography_vector = 67e6,
  initial_conditions = matrix(
    c(0.9999, 0, 0, 0, 0, 0.0001, 0, 0, 0, 0, 0),
    nrow = 1, ncol = 11L
  )
)

# create a vaccination regime
double_vax <- vaccination(
  nu = matrix(1e-3, ncol = 2, nrow = 1),
  time_begin = matrix(c(10, 30), nrow = 1),
  time_end = matrix(c(50, 80), nrow = 1)
)

# run epidemic simulation with no vaccination or intervention
output <- model_vacamole_cpp(
  population = population,
  vaccination = double_vax
)

# view the output
output
}
\references{
Ainslie, K. E. C., Backer, J. A., Boer, P. T. de, Hoek, A. J. van,
Klinkenberg, D., Altes, H. K., Leung, K. Y., Melker, H. de, Miura, F., &
Wallinga, J. (2022). A scenario modelling analysis to anticipate the impact
of COVID-19 vaccination in adolescents and children on disease outcomes in
the Netherlands, summer 2021. Eurosurveillance, 27(44), 2101090.
\doi{10.2807/1560-7917.ES.2022.27.44.2101090}
}
