% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{new_infections}
\alias{new_infections}
\title{Get new infections}
\usage{
new_infections(
  data,
  compartments_from_susceptible = "vaccinated",
  by_group = TRUE
)
}
\arguments{
\item{by_group}{}
}
\description{
Get new infections
}
