% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{no_intervention}
\alias{no_intervention}
\title{Generate a null intervention}
\usage{
no_intervention(population)
}
\arguments{
\item{population}{A \code{population} object with a \code{contact_matrix} member.}
}
\value{
An intervention that has no effect on contacts, with start and end
times set to 0.0
}
\description{
Generate a null intervention
}
