% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_dependence.R
\name{no_time_dependence}
\alias{no_time_dependence}
\title{Specify no time dependence of model rates}
\usage{
no_time_dependence()
}
\value{
A list with a single named element, \code{beta}, which is a function
that returns its second argument. This matches the specification for
functions to be passed as the \code{time_dependence} argument of epidemic
functions.
}
\description{
Specify no time dependence of model rates
}
