% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{no_vaccination}
\alias{no_vaccination}
\title{Generate a null vaccination}
\usage{
no_vaccination(population)
}
\arguments{
\item{population}{A \code{population} object with a \code{contact_matrix} member.}
}
\value{
An vaccination that has no effect on the population, with start and
end times set to 0.0, and the rate of vaccination \eqn{nu} also set to 0.0.
}
\description{
Generate a null vaccination
}
