% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Non-Centered_MCMC.R
\name{NonCentered_MCMC}
\alias{NonCentered_MCMC}
\title{MCMC Sampler for a non-centered Bayesian Parameterisation of
Epidemics}
\usage{
NonCentered_MCMC(
  N,
  a,
  t_rem,
  gamma0,
  theta_gamma,
  beta0,
  theta_beta,
  kernel,
  no_its,
  burn_in = 0,
  no_proposals,
  lambda,
  PLOT = TRUE,
  thinning_factor = 1,
  lag_max = NA
)
}
\description{
MCMC Sampler for a non-centered Bayesian Parameterisation of
Epidemics
}
