% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneParticlePM_SIS_beta.R
\name{oneParticlePM_SIS_beta}
\alias{oneParticlePM_SIS_beta}
\title{Create Storage Matrix
Proposal Acceptance Counter
Propose new beta and gamma using Multiplicative RW propsal
Store State
Calculating Summary Statistics for samples}
\usage{
oneParticlePM_SIS_beta(
  obsTransData,
  obsTimes,
  N,
  beta0,
  gamma,
  lambda,
  noIts,
  burnIn = 0,
  lagMax = NA,
  thinningFactor = 1
)
}
\description{
Create Storage Matrix
Proposal Acceptance Counter
Propose new beta and gamma using Multiplicative RW propsal
Store State
Calculating Summary Statistics for samples
}
