% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_data.R
\name{panel_data_new.epidemics}
\alias{panel_data_new.epidemics}
\title{Temporary Panel Data Generator}
\usage{
panel_data_new.epidemics(
  subject,
  times,
  state,
  prev_state,
  trans = possible_transitions(1:3),
  subset = NULL,
  obs_times
)
}
\description{
Panel Data
}
