% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_gillespies.R
\name{panelDataSIR_Gillespie}
\alias{panelDataSIR_Gillespie}
\title{Possible states and their respective
logical representations.
Summarise in matrix
Initial state of each individual
Initial state of epidemic at population level
Possible Events
Summarise in a matrix
Panel Data
Waiting Time
Panel Data
Convert individualStates into panelData
currentState = sapply(X = 1:N, function(X) which(individualStates[X,]))
Which event happens?
To who?
Update Population State
Update Individual State
Convert individualStates into panelData}
\usage{
panelDataSIR_Gillespie(
  initialState,
  beta,
  gamma,
  obsTimes,
  kernel,
  replace = F
)
}
\description{
Possible states and their respective
logical representations.
Summarise in matrix
Initial state of each individual
Initial state of epidemic at population level
Possible Events
Summarise in a matrix
Panel Data
Waiting Time
Panel Data
Convert individualStates into panelData
currentState = sapply(X = 1:N, function(X) which(individualStates[X,]))
Which event happens?
To who?
Update Population State
Update Individual State
Convert individualStates into panelData
}
