% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Psuedo_Marginal_MCMC.R
\name{pi_Y_given_X}
\alias{pi_Y_given_X}
\title{Psuedo-Marginal MCMC For Inference on Epidemic Data}
\usage{
pi_Y_given_X(N, Y, obs_times, initial_infective, beta, gamma)
}
\arguments{
\item{N, }{size of complete population}

\item{Y, }{observed subset of the complete epidemic data}

\item{k, }{noumber of observations of epidemic}

\item{T_obs, }{amount of time epidemic is observed for.}
}
\description{
Takes multiple discrete observations of an epidemic
and the complete observation and evaluates the
pdf of observing the partial information given
the complete information.
function to simulate an epidemic with parameters theta and calculate pi(y|x)
}
