% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.epidemic.R
\name{plot.epidemic}
\alias{plot.epidemic}
\title{Function for plotting the evolution of epidemic}
\usage{
\method{plot}{epidemic}(times, N, X, Y, Z, newPlot = TRUE)
}
\description{
Function for plotting the evolution of epidemic
}
