% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM-MCMCpanelExperiment.R
\name{PM_MCMCexperiment}
\alias{PM_MCMCexperiment}
\title{PM-MCMC Panel Experiment
Sample Panel Data}
\usage{
PM_MCMCexperiment(
  noSims,
  noPanels,
  lastObs,
  lambda0,
  V0 = diag(c(1/N, 1)),
  adapt = TRUE,
  noIts
)
}
\description{
PM-MCMC Panel Experiment
Sample Panel Data
}
