% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partially_ Non-Centered_MCMC.R
\name{PNC.Algorithm}
\alias{PNC.Algorithm}
\title{PhD
Partially Non-Centered Sampling Algorithm for Epidemics}
\usage{
PNC.Algorithm(
  N,
  a,
  t_rem,
  gamma0,
  beta0,
  theta.gamma,
  theta.beta,
  rinf.dist,
  no_proposals,
  lambda,
  centering_prob,
  no_its,
  burn_in,
  thinning_factor = 1,
  lag_max = NA
)
}
\description{
PhD
Partially Non-Centered Sampling Algorithm for Epidemics
}
