% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_default.R
\name{prepare_args_default}
\alias{prepare_args_default}
\title{Prepare arguments to default epidemic function}
\usage{
prepare_args_default(mod_args)
}
\description{
Prepare arguments to default epidemic function
}
\keyword{internal}
