% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_default.R
\name{prepare_default_args}
\alias{prepare_default_args}
\alias{.check_prepare_args_default}
\title{Prepare arguments to default model function}
\usage{
.check_prepare_args_default(mod_args)
}
\arguments{
\item{mod_args}{A named list of the population, and epidemic modifiers.}
}
\value{
A list of model arguments suitable for .model_default_cpp().
This is a named list consisting of:
\itemize{
\item \code{initial_state}: the initial conditions modified to represent absolute
rather than proportional values;
\item \code{transmission_rate}, \code{infectiousness_rate}, \code{recovery_rate}: three numbers
representing the transmission rate of the infection, the rate of transition
from exposed to infectious, and the recovery rate, respectively;
\item \code{contact_matrix}, a numeric matrix for the population contact matrix
scaled by the largest real eigenvalue and by the size of each groups;
\item \code{npi_time_begin}, \code{npi_time_end}: two vectors for the start and end times
of any interventions applied;
\item \code{npi_cr}: a matrix for the age- and intervention-specific effect on social
contacts;
\item \code{vax_time_begin},\code{vax_time_end}, \code{vax_nu}: three numeric matrices for the
age- and dose-specific start times, end times, and rates of any vaccination
doses implemented;
\item \code{time_end}, \code{increment}: two numbers for the time at which to end the
simulation, and the value by which the simulation time
is incremented.
}
}
\description{
Prepare arguments to \code{\link[=model_default]{model_default()}} for
.model_default_cpp().
}
\details{
\code{.check_prepare_args_default()} prepares arguments for
.model_default_cpp(), which is the C++ function that solves the default
ODE system using a Boost \emph{odeint} solver, by converting some of the arguments
collected in \code{mod_args} into simpler structures such as lists and numeric or
integer vectors that can be interpreted as C++ types such as \code{Rcpp::List},
\code{Rcpp::NumericVector}, or \code{Eigen::MatrixXd}.
}
\keyword{internal}
