% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_diphtheria.R
\name{prepare_diphtheria_args}
\alias{prepare_diphtheria_args}
\alias{.check_prepare_args_diphtheria}
\title{Prepare arguments to diphtheria model function}
\usage{
.check_prepare_args_diphtheria(mod_args)
}
\arguments{
\item{mod_args}{A named list of the population, and epidemic modifiers.}
}
\value{
A list of model arguments suitable for model_diphtheria().
This is a named list consisting of:
\itemize{
\item \code{initial_state}: the initial conditions modified to represent absolute
rather than proportional values;
\item \code{transmission_rate}, \code{transmission_rate_vax}: two numbers representing the
transmission rate
of the infection for unvaccinated or single-dose vaccinated, and two-dose
vaccinated individuals, respectively;
\item \code{infectiousness_rate}: a single number for the transition rate from the
'exposed' and 'exposed_vaccinated' to the 'infectious' and
'infectious_vaccinated' compartments;
\item \code{recovery_rate}: a single number for the recovery rate from the infection;
\item \code{reporting_rate}: a single number for the proportion of infectious cases
reported;
\item \code{prop_hosp}: a single number for the proportion of reported cases that
need hospitalisation;
\item \code{hosp_entry_rate}, \code{hosp_exit_rate}: two numbers representing the rate of
entry and exit from the 'hospitalised' compartment;
\item \code{rate_interventions}: a List giving the interventions on model
parameters;
\item \code{time_dependence}: a List giving the time-dependent effects on model
parameters in the form of R functions;
\item \code{pop_change_times} and \code{pop_change_values}: the times and values of changes
in the population of susceptibles;
\item \code{time_end}, \code{increment}: two numbers for the time at which to end the
simulation, and the value by which the simulation time
is incremented.
}
}
\description{
Prepare arguments to \code{\link[=model_diphtheria]{model_diphtheria()}}.
}
\details{
\code{.check_prepare_args_diphtheria()} prepares arguments for
model_diphtheria(), which uses an odin C function to solve the
ODE system.
}
\keyword{internal}
