% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_ebola.R
\name{prepare_ebola_args}
\alias{prepare_ebola_args}
\alias{.check_prepare_args_ebola}
\title{Prepare arguments to ebola model function}
\usage{
.check_prepare_args_ebola(mod_args)
}
\arguments{
\item{mod_args}{A named list of composable elements; must include a
\verb{<population>}.}
}
\value{
A list of model arguments suitable for \code{\link[=.model_ebola_internal]{.model_ebola_internal()}}.
This is a named list consisting of:

\strong{WIP}
}
\description{
Prepare arguments to \code{\link[=model_ebola]{model_ebola()}} for
\code{\link[=.model_ebola_internal]{.model_ebola_internal()}}.
}
\details{
\code{.check_prepare_args_ebola()} prepares arguments for
\code{\link[=.model_ebola_internal]{.model_ebola_internal()}}, by converting some of the arguments
collected in \code{mod_args} into simpler structures that are appropriate for the
internal model function.
}
\keyword{internal}
