% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_vacamole.R
\name{prepare_vacamole_args}
\alias{prepare_vacamole_args}
\alias{.check_prepare_args_vacamole}
\title{Prepare arguments to Vacamole epidemic function}
\usage{
.check_prepare_args_vacamole(mod_args)
}
\value{
A list of model arguments suitable for .model_vacamole_cpp().
This is a named list consisting of:
\itemize{
\item \code{initial_state}: the initial conditions modified to represent absolute
rather than proportional values;
\item \code{transmission_rate}, \code{transmission_rate_vax}: two numbers representing the
transmission rate
of the infection for unvaccinated or single-dose vaccinated, and two-dose
vaccinated individuals, respectively;
\item \code{infectiousness_rate}: a single number for the transition rate from the
'exposed' and 'exposed_vaccinated' to the 'infectious' and
'infectious_vaccinated' compartments;
\item \code{mortality_rate}, \code{mortality_rate_vax}: two numbers representing the
mortality rate of the infection for unvaccinated or single-dose vaccinated,
and two-dose vaccinated individuals, respectively;
\item \code{hospitalisation_rate}, \code{hospitalisation_rate_vax}: two numbers
representing the hospitalisation rate of the infection for unvaccinated or
single-dose vaccinated, and two-dose vaccinated individuals, respectively;
\item \code{recovery_rate}: a single number for the recovery rate from the infection;
\item \code{contact_matrix}, a numeric matrix for the population contact matrix
scaled by the largest real eigenvalue and by the size of each groups;
\item \code{npi_time_begin}, \code{npi_time_end}: two vectors for the start and end times
of any interventions applied;
\item \code{npi_cr}: a matrix for the age- and intervention-specific effect on social
contacts;
\item \code{vax_time_begin},\code{vax_time_end}, \code{vax_nu}: three numeric matrices for the
age- and dose-specific start times, end times, and rates of any vaccination
doses implemented;
\item \code{time_end}, \code{increment}: two numbers for the time at which to end the
simulation, and the value by which the simulation time
is incremented.
}
}
\description{
Prepare arguments to \code{\link[=model_vacamole]{model_vacamole()}} for
.model_vacamole_cpp().
}
\details{
\code{.check_prepare_args_vacamole()} prepares arguments for
.model_vacamole_cpp(), which is the C++ function that solves the Vacamole
ODE system using a Boost \emph{odeint} solver, by converting some of the arguments
collected in \code{mod_args} into simpler structures such as lists and numeric or
integer vectors that can be interpreted as C++ types such as \code{Rcpp::List},
\code{Rcpp::NumericVector}, or \code{Eigen::MatrixXd}.
}
\keyword{internal}
