% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection.R
\name{print.infection}
\alias{print.infection}
\title{Print an \verb{<infection>} object}
\usage{
\method{print}{infection}(x, ...)
}
\arguments{
\item{x}{A \verb{<infection>} object.}

\item{...}{Other parameters passed to \code{\link[=print]{print()}}.}
}
\value{
Invisibly returns the \verb{<infection>} object \code{x}.
Called for printing side-effects.
}
\description{
Print an \verb{<infection>} object
}
