% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{print.intervention}
\alias{print.intervention}
\title{Print a \code{intervention} object}
\usage{
\method{print}{intervention}(x, ...)
}
\arguments{
\item{x}{A \code{intervention} object.}

\item{...}{Other parameters passed to \code{\link[=print]{print()}}.}
}
\value{
None. Prints output.
}
\description{
Print a \code{intervention} object
}
