% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{print.population}
\alias{print.population}
\title{Print a \verb{<population>} object}
\usage{
\method{print}{population}(x, ...)
}
\arguments{
\item{x}{A \verb{<population>} object.}

\item{...}{Other parameters passed to \code{\link[=print]{print()}}.}
}
\value{
Invisibly returns the \verb{<population>} object \code{x}.
Called for printing side-effects.
}
\description{
Print a \verb{<population>} object
}
