% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{print.vaccination}
\alias{print.vaccination}
\title{Print a \verb{<vaccination>} object}
\usage{
\method{print}{vaccination}(x, ...)
}
\arguments{
\item{x}{A \verb{<vaccination>} object.}

\item{...}{Other parameters passed to \code{\link[=print]{print()}}.}
}
\value{
Invisibly returns the \verb{<vaccination>} object \code{x}.
Called for printing side-effects.
}
\description{
Print a \verb{<vaccination>} object
}
