% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_class.R
\name{print_epidemic}
\alias{print_epidemic}
\alias{print.epidemic}
\title{Print an object of the \verb{<epidemic>} class}
\usage{
\method{print}{epidemic}(x, ...)
}
\arguments{
\item{x}{An \verb{<epidemic>} class object.}

\item{...}{Other parameters passed to \code{\link[=print]{print()}}.}
}
\value{
Invisibly returns the object \code{x}. Called for printing side-effects.
}
\description{
Print an object of the \verb{<epidemic>} class
}
