% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_Calculations.R
\name{prod_part_inf}
\alias{prod_part_inf}
\title{Infection component of Heterogeneous Likelihood Functions
Original Author: Chris Jewell (2019)
Link: http://fhm-chicas-code.lancs.ac.uk/jewell/epilikelihoods/tree/master/R}
\usage{
prod_part_inf(t_inf_j, events, B, log = TRUE)
}
\description{
Infection component of Heterogeneous Likelihood Functions
Original Author: Chris Jewell (2019)
Link: http://fhm-chicas-code.lancs.ac.uk/jewell/epilikelihoods/tree/master/R
}
