% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Psuedo_Marginal_MCMC.R
\name{Pseudo_Marginal_MCMC}
\alias{Pseudo_Marginal_MCMC}
\title{Variable List to Pass to Cores
Intialise Missing Data by Simulating no_sims epidemics with
parameters beta and gamma
Create Storage Matrix
Proposal Acceptance Counter
Propose new beta and gamma using Multiplicative RW propsal
Store State
Calculating Summary Statistics for samples}
\usage{
Pseudo_Marginal_MCMC(
  N,
  Y,
  obs_times,
  beta0,
  gamma0,
  prior_rate,
  initial_infective,
  lambda,
  V,
  no_sims,
  no_its,
  burn_in,
  lag_max = NA,
  thinning_factor = 1,
  parallel = FALSE,
  no_cores
)
}
\description{
Variable List to Pass to Cores
Intialise Missing Data by Simulating no_sims epidemics with
parameters beta and gamma
Create Storage Matrix
Proposal Acceptance Counter
Propose new beta and gamma using Multiplicative RW propsal
Store State
Calculating Summary Statistics for samples
}
