% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_library.R
\name{read_from_library}
\alias{read_from_library}
\title{Get functions from model library}
\usage{
read_from_library(
  model_type = "epidemic",
  model_name = "default",
  which_function = "model_function"
)
}
\arguments{
\item{model_type}{The model type. The only option currently available is
"epidemic". Future development may see the addition of more model types,
including endemic disease compartmental models, and branching process
outbreak models.}

\item{model_name}{The model name. The only option currently available is
"default", for the default epidemic compartmental SEIR-V model.}

\item{which_function}{A string for the function requested, which may be one
of "model_function", "model_args_checker", and "model_args_prepper". These
correspond to functions that run the model, check its arguments, and prepare
the arguments for the model function, respectively.}
}
\description{
Get functions from model library
}
\keyword{internal}
