% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Epidemic_Simulation.R
\name{relevant_rates}
\alias{relevant_rates}
\title{Seems like drawing the time to next event and which event will occur will use
the same information, but in different ways. Either way, both functions will
have to weedle out which rates are relevant given what state people are in.
We do not want to have to decide this more than once, therefore there should
be a function which does this, and then we can pass the result to the functions.}
\usage{
relevant_rates(rates, population)
}
\arguments{
\item{rates}{Holds information about the rates ate which each individual transition
between each state}

\item{population}{Holds information about what state each individual in the population
is in}
}
\value{
relevant_rates
}
\description{
relevant_rates A function which will decided which rates are important, in terms of
               deciding when the next event occurs and which event occurs. The aim of
               this function is to avoid extracting this information twice.
}
