% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWMEpidemics.R
\name{RWMSampler}
\alias{RWMSampler}
\title{(Multiplicative) Random-Walk Metropolis for Heterogeneous SIR Epidemic}
\usage{
RWMSampler(
  beta0,
  betaPrior = c(1, 0.001),
  gammaPrior = c(1, 0.001),
  N,
  eventTimes,
  kernel,
  lambda,
  V = 1,
  noIts,
  burnIn,
  thinFactor = 1,
  lagMax = NULL
)
}
\description{
Full information on epidemic
}
\details{
Heterogeneous Kernel set up

Infectious model is parameterised by a predefined kernel.

Removal process is independent of infectious process a posteriori

Hence no need to sample removal parameter,
given we have chosen a suitable prior (i.e Exponential/Gamma)

Paired Infection-Removal times.

Initialise
Calculate Infectious Process part of the likelihood (Removal Process will cancel out).
Gamma Posterior Parameters
Data
Accept/Reject
MCMC Plots
}
