% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_desolve.R
\name{sir_desolve}
\alias{sir_desolve}
\title{SIR epidemic model using 'deSolve'}
\usage{
sir_desolve(beta, gamma, S, I, R, times)
}
\arguments{
\item{beta}{Transmission rate \eqn{beta}.}

\item{gamma}{Recovery rate \eqn{gamma}.}

\item{S}{The proportion of individuals initially susceptible to infection.}

\item{I}{The proportion of individuals initially infected by the pathogen.}

\item{R}{The proportion of individuals already recovered from infection
which cannot be infected again.}

\item{times}{A sequence of timepoints at which to solve for the proportion of}
}
\value{
A list with the proportion of the population in the
susceptible, infected, and recovered states at each time point in \code{times}.
}
\description{
SIR epidemic model using 'deSolve'
}
\examples{
data <- sir_desolve(
  beta = 0.1, gamma = 0.05, S = 0.99, I = 0.01, R = 0.0,
  times = seq(0, 100, 0.1)
)

str(data)
}
