% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockSamplesSIRfsMCMC.R
\name{SIR_fsMCMC_blockedIS}
\alias{SIR_fsMCMC_blockedIS}
\title{fsMCMC MCMC for SIR Epidemic Panel Data (with blocked proposals of underlying Random Variables)}
\usage{
SIR_fsMCMC_blockedIS(
  obsTransData,
  I_0,
  obsTimes,
  N,
  beta0,
  gamma0,
  lambda,
  V,
  noDraws = 2 * N - I_0,
  blockSize,
  noIts,
  burnIn = 0,
  lagMax = NA,
  thinningFactor = 1
)
}
\arguments{
\item{obsTransData}{Interpanel transition data.}

\item{I_0}{Initial number of infectives in the population.}

\item{obsTimes}{Times at which epidemic cohort were followed up.}

\item{N}{Population size.}

\item{beta0}{Starting value for infectious process parameter.}

\item{gamma0}{Starting value for removal/recovery process parameter.}

\item{lambda}{Starting value for RWM proposal parameter which is to be adapted.}

\item{V}{Starting state for RWM proposal Covariance matrix which is to be adapted.}

\item{noDraws}{Number of underlying variables to be drawn (i.e max. number of events in observation period)}

\item{blockSize}{Number of underlying random variables of the process to refresh.}

\item{noIts}{Number of MCMC iterations.}

\item{lagMax}{Plotting parameter for acf() function.}

\item{thinningFactor}{Controls the factor by which MCMC samples are thinned, to reduce dependency.}
}
\value{
MCMC summary.
}
\description{
Adapts proposal parameters for with a view of optimal of a target using forward simulation MCMC scheme.
}
\seealso{
Other Panel Data MCMC: 
\code{\link{adaptiveSIR_PseudoMarginalMCMC}()},
\code{\link{adaptiveSIS_PseudoMarginalMCMC}()},
\code{\link{adaptiveSISfsMCMC}()}
}
\concept{Panel Data MCMC}
