% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_Gillespie.R
\name{SIR_Gillespie}
\alias{SIR_Gillespie}
\title{SIR_Gillespie function}
\usage{
SIR_Gillespie(
  N,
  initial_state = "start",
  beta,
  gamma,
  trans = possible_transitions(1:3),
  obs_end = Inf,
  kernel = NULL,
  U,
  E,
  output = "event",
  obs_times = NULL
)
}
\description{
Takes parameters for an SIR Infectious Process and returns either an event table or
panel data
Initial State, a vector of length N stating who is which infectious state and
the current time.
Track specific people
Infection Matrix
Last and unused panels
}
