% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_PM-MCMCexperiment1.R
\name{SIR_PM_MCMCexperiment1}
\alias{SIR_PM_MCMCexperiment1}
\title{SIR Test Epidemic 1 PM-MCMC Experiment
Test data}
\usage{
SIR_PM_MCMCexperiment1(noPanels, noSims, lambda0, V0, adapt = T, noIts = 1e+06)
}
\description{
SIR Test Epidemic 1 PM-MCMC Experiment
Test data
}
