% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_PM-MCMC.R
\name{SIR_PseudoMarginalMCMC}
\alias{SIR_PseudoMarginalMCMC}
\title{SIR PM-MCMC
Create Storage Matrix
Proposal Acceptance Counter
Propose new beta and gamma using Multiplicative RW propsal
Folded Normal
Store State
Calculating Summary Statistics for samples}
\usage{
SIR_PseudoMarginalMCMC(
  obsTransData,
  I_0,
  obsTimes,
  N,
  beta0,
  gamma0,
  lambda,
  V,
  noSims,
  noIts,
  burnIn,
  lagMax = NA,
  thinningFactor = 1,
  parallel = FALSE,
  noCores
)
}
\description{
SIR PM-MCMC
Create Storage Matrix
Proposal Acceptance Counter
Propose new beta and gamma using Multiplicative RW propsal
Folded Normal
Store State
Calculating Summary Statistics for samples
}
