% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRfsMCMCexperiment.R
\name{SIRfsMCMCexperiment1}
\alias{SIRfsMCMCexperiment1}
\title{Testing New blocked Independent Sampler fsMCMC for SIR (+ Adaptive Step)
SIR fsMCMC experiment
Multiple Datasets, Different speed outbreaks.
Test data}
\usage{
SIRfsMCMCexperiment1(
  noPanels,
  lastObs,
  lambda0 = 0.001,
  V0,
  blockSize,
  adapt = T,
  noIts = 1e+06
)
}
\description{
Testing New blocked Independent Sampler fsMCMC for SIR (+ Adaptive Step)
SIR fsMCMC experiment
Multiple Datasets, Different speed outbreaks.
Test data
}
