% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIS_fsMCMC_new.R
\name{SIS_fsMCMC_beta}
\alias{SIS_fsMCMC_beta}
\title{Create Storage Matrix
Proposal Acceptance Counter
Folded Normal
Store State
Calculating Summary Statistics for samples}
\usage{
SIS_fsMCMC_beta(
  obsTransData,
  I_0,
  obsTimes,
  N,
  beta0,
  gamma,
  thetaLim,
  lambda,
  noDraws,
  s,
  noIts,
  burnIn = 0,
  lagMax = NA,
  thinningFactor = 1
)
}
\description{
Create Storage Matrix
Proposal Acceptance Counter
Folded Normal
Store State
Calculating Summary Statistics for samples
}
