% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SISfsMCMC_blockedIndependenceSampler.R
\name{SIS_fsMCMC_blockedIS}
\alias{SIS_fsMCMC_blockedIS}
\title{Calculate the amount of blockSamples which need to be made
To accomodate for a potentially different final block size (blockSamples is not integer), calculate the residual noDraws
Create Storage Matrix
Proposal Acceptance Counter
Folded Normal
Store State
Calculating Summary Statistics for samples}
\usage{
SIS_fsMCMC_blockedIS(
  obsTransData,
  I_0,
  obsTimes,
  N,
  beta0,
  gamma0,
  thetaLim,
  lambda,
  V,
  noDraws,
  blockSize,
  noIts,
  burnIn = 0,
  lagMax = NA,
  thinningFactor = 1
)
}
\description{
Calculate the amount of blockSamples which need to be made
To accomodate for a potentially different final block size (blockSamples is not integer), calculate the residual noDraws
Create Storage Matrix
Proposal Acceptance Counter
Folded Normal
Store State
Calculating Summary Statistics for samples
}
