% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIS_Gillespie.R
\name{SIS_Gillespie}
\alias{SIS_Gillespie}
\title{SIS_Gillespie function}
\usage{
SIS_Gillespie(
  N,
  a,
  beta,
  gamma,
  trans = possible_transitions(1:2),
  obs_end = Inf,
  kernel = NULL,
  U = NULL,
  E = NULL,
  output = "event",
  obs_times = NULL
)
}
\description{
Takes parameters for an SIR Infectious Process and returns either an event table or
panel data.
Track specific people
Infection Matrix
Last and unused panels
}
