% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIS_PM-MCMC.R
\name{SIS_PseudoMarginalMCMC}
\alias{SIS_PseudoMarginalMCMC}
\title{New Pseudo-Marginal SIS
Create Storage Matrix
Proposal Acceptance Counter
Propose new beta and gamma using Multiplicative RW propsal
Folded Normal
Store State
Calculating Summary Statistics for samples}
\usage{
SIS_PseudoMarginalMCMC(
  obsTransData,
  I_0,
  obsTimes,
  N,
  beta0,
  gamma0,
  lambda,
  V,
  noSims,
  noIts,
  burnIn,
  lagMax = NA,
  thinningFactor = 1,
  parallel = FALSE,
  noCores
)
}
\description{
New Pseudo-Marginal SIS
Create Storage Matrix
Proposal Acceptance Counter
Propose new beta and gamma using Multiplicative RW propsal
Folded Normal
Store State
Calculating Summary Statistics for samples
}
