% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SISpanelExperiment.R
\name{SISfsMCMCexperiment}
\alias{SISfsMCMCexperiment}
\title{Variance of beta, gamma for varying number of panels.}
\usage{
SISfsMCMCexperiment(
  noPanels,
  lastObs,
  noDraws,
  lambda0,
  V0,
  blockSize,
  adapt = TRUE,
  noIts = 10000,
  burnIn = 1000
)
}
\description{
Same SIS epidemic which occurs in population size N, (set.seed(1))
}
\details{
We observe the same m = 0.1*N individuals, but vary how often they are observed.

In this simulated situation, the average infectious period is a unit length of time (\gamma = 1).

R_0 is chosen to be above 1 and relitively small, but large enough that the disease will reach an endemic state.

\beta is then calculated from \gamma, R_0 and N

Sample Panel Data
}
