% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_data_epidemics.R
\name{start_state}
\alias{start_state}
\title{Transition Panel Data for a General Epidemic Event Table}
\usage{
start_state(subject, prev_state)
}
\description{
Function aim:
}
\details{
Take a event table, which is output from an Epidemic Gillespie algorithm,
and return panel data on the number of each possible transitions that happened
between the specified observation points.

An event table will have a few things. For each event, there will be ...


 Time of event
 ID of individual
 State of Individual after event
 State of Individual before event
 Covariates of individual

Then we can give panel_data.epidemics the event table
perhaps tell it a subset of individuals we want to observe
also, and got transition panel data out.

function to keep transitions with zero transition count in table
}
