% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_table.R
\name{state_table}
\alias{state_table}
\title{state_table function}
\usage{
state_table(x, y, trans)
}
\description{
Takes two vectors of states and the possible transitions,
then returns counts of unique transitions
}
