% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{test_recyclable}
\alias{test_recyclable}
\title{Check compliance with Tidyverse vector recycling rules}
\usage{
test_recyclable(x)
}
\arguments{
\item{x}{A list of vectors to be checked for compliance with Tidyverse
recycling rules.}
}
\value{
A single logical value.
}
\description{
Check compliance with Tidyverse vector recycling rules
}
\author{
Tim Taylor
}
\keyword{internal}
