% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionData.R
\name{transitionData}
\alias{transitionData}
\title{Function which transforms panel data into Transition Data}
\usage{
transitionData(panelData, states)
}
\description{
Function which transforms panel data into Transition Data
}
