% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vaccinated}
\alias{vaccinated}
\title{Get the number of people vaccinated}
\usage{
vaccinated(data)
}
\arguments{
\item{data}{A \verb{<data.frame>} of model output, typically
the output of a compartmental model.}
}
\value{
A \verb{<data.frame>} with columns "demography_group", "vaccinated",
"total" and "rescale"; these specify the name of the demography group,
the number of vaccinated, the total number of people in each demography
group and the rescale factor.
}
\description{
Gets the number of people vaccinated in each demography group and
calculate a rescale factor which when multiplied with \eqn{\nu} returns
the effective vaccination rate.
}
\examples{
polymod <- socialmixr::polymod
contact_data <- socialmixr::contact_matrix(
  polymod,
  countries = "United Kingdom",
  age.limits = c(0, 20, 65),
  symmetric = TRUE
)

# prepare contact matrix
contact_matrix <- t(contact_data$matrix)

# prepare the demography vector
demography_vector <- contact_data$demography$population
names(demography_vector) <- rownames(contact_matrix)
# initial conditions
initial_i <- 1e-6
initial_conditions <- c(
  S = 1 - initial_i, E = 0, I = initial_i, R = 0, V = 0
)

# build for all age groups
initial_conditions <- rbind(
  initial_conditions,
  initial_conditions,
  initial_conditions
)

# assign rownames for clarity
rownames(initial_conditions) <- rownames(contact_matrix)

uk_population <- population(
  name = "UK",
  contact_matrix = contact_matrix,
  demography_vector = demography_vector,
  initial_conditions = initial_conditions
)

# prepare a vaccination object
vaccinate_elders <- vaccination(
  name = "vaccinate elders",
  time_begin = matrix(100, nrow(contact_matrix)),
  time_end = matrix(250, nrow(contact_matrix)),
  nu = matrix(c(0, 0, 0.0001))
)
# run an epidemic model using `epidemic`
output <- model_default_cpp(
  population = uk_population,
  vaccination = vaccinate_elders,
  time_end = 600, increment = 1.0
)

# get the number of vaccinated and rescale factor for each demography group
num_vaccinated <- vaccinated(output)
num_vaccinated
}
