% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_class.R
\name{validate_epidemic}
\alias{validate_epidemic}
\title{Validate \if{html}{\out{<epidemic>}} class objects}
\usage{
validate_epidemic(x)
}
\arguments{
\item{x}{An \verb{<epidemic>} object.}
}
\value{
Invisibly returns the input \code{x}. Called primarily for its
input checking as a validator.
}
\description{
Validate \if{html}{\out{<epidemic>}} class objects
}
\keyword{internal}
