% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{validate_intervention}
\alias{validate_intervention}
\alias{validate_contacts_intervention}
\alias{validate_rate_intervention}
\title{Validate objects that inherit from the \if{html}{\out{<intervention>}} class}
\usage{
validate_contacts_intervention(x)

validate_rate_intervention(x)
}
\arguments{
\item{x}{For \code{validate_contacts_intervention()}, an object to be validated as
a \verb{<contacts_intervention>}. For \code{validate_rate_intervention()}, an object to
be validated as a \verb{<rate_intervention>}.}
}
\value{
Invisibly returns the input \code{x}. Called primarily for its
input checking as a validator for objects of the \verb{<intervention>} superclass.
}
\description{
Validate objects that inherit from the \if{html}{\out{<intervention>}} class

Validate objects that inherit from the \if{html}{\out{<intervention>}} class
}
\keyword{internal}
