% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{vector_recycling}
\alias{vector_recycling}
\alias{.test_recyclable}
\alias{.recycle_vectors}
\title{Vector recyclability checks and vector recycling}
\usage{
.test_recyclable(x)

.recycle_vectors(x)
}
\arguments{
\item{x}{A list of vectors to be checked for compliance with Tidyverse
recycling rules, or to be recycled.
When \code{x} is a list to be recycled, two cases are envisaged: all
elements of \code{x} are expected to be of the same length, or, all but one
element of \code{x} are scalars.}
}
\value{
\code{.test_recyclable} returns a single logical value.
\code{.recycle_vectors} returns a list of vectors recycled to the same length
following Tidyverse recycling rules.
}
\description{
Internal functions to check whether vectors can be recycled and
to recycle vectors.
}
\details{
Note that \code{.test_recyclable()} will return vectors of unequal lengths if \code{x}
does not comply with length rules. This compliance is not enforced as this
function is only expected to be used internally, after a call to
\code{.test_recyclable()} (e.g. in a \code{stopifnot()}).
}
\author{
Tim Taylor
}
\keyword{internal}
