% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Epidemic_Simulation.R
\name{which_event}
\alias{which_event}
\title{which_event Decides which event will occur next in the Infectious Disease process}
\usage{
which_event(rates, population)
}
\arguments{
\item{rates}{Holds information about the rates ate which each individual transition
between each state.}

\item{population}{Holds information about what state each individual in the
population is in.}
}
\description{
which_event Decides which event will occur next in the Infectious Disease process
}
