% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{as.intervention}
\alias{as.intervention}
\title{Convert a list to a intervention object}
\usage{
as.intervention(x, type = c("contacts", "rate"))
}
\arguments{
\item{x}{A list, or an object that inherits from a list.}

\item{type}{A string for the type of intervention: \code{"contacts"} for a
\verb{<contact_intervention>} or \code{"rate"} for a \verb{<rate_intervention>}.}
}
\value{
A \link{intervention} class object.
}
\description{
Convert a list to a intervention object
}
\examples{
# prepare a list
npi <- list(
  name = "npi",
  type = "contacts",
  time_begin = 30,
  time_end = 60,
  reduction = rep(0.1, 3)
)

as.intervention(npi)
}
