% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check_helpers.R
\name{assert_infection}
\alias{assert_infection}
\title{Assert properties of an \code{infection} object}
\usage{
assert_infection(
  x,
  default_params = c("name", "r0", "infectious_period"),
  extra_parameters = NULL,
  extra_parameters_limits = NULL
)
}
\arguments{
\item{x}{An \link{infection} object.}

\item{default_params}{Default parameter names present in an
\code{infection} object.
This argument is provided to account for potential changes to the default
elements of an \code{infection} object, and is not expected to be changed.}

\item{extra_parameters}{A character vector giving the names of any extra
parameters included in \code{x}. \code{x} may already be expected to have the members
\code{r0} and \code{infectious_period}.}

\item{extra_parameters_limits}{An optional named list of the numeric limits
of the acceptable limits of the parameters specifies in \code{extra_parameters}.
Each list element must be a named, two element vector of 'lower' and 'upper'
limits. Not all extra parameters need to have their limits passed explicitly.
Parameters whose limits are not explicitly specified will have their values
checked against a lower bound of 0, with the upper limit expected to be
finite.}
}
\value{
Silently returns the \code{infection} object \code{x}. Primarily called for its
side effects of throwing errors when \code{x} does not meet certain requirements.
}
\description{
Assert that objects of the \code{infection} class have the parameters expected by
an epidemic model. See \code{\link[=infection]{infection()}} and specific epidemic functions,
to check the infection parameters required by each model. This
function is for internal use in argument checking functions.
}
\examples{
\dontrun{
# prepare a well formed infection object for the default model
infection_default <- infection(
  name = "influenza", r0 = 1.3, infectious_period = 10,
  preinfectious_period = 3, mortality_rate = 1e-4
)

# expect no errors for well formed infection-assertion matches
assert_infection(
  infection_default,
  extra_parameters = c("preinfectious_period", "mortality_rate")
)
}
}
\keyword{internal}
