% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_default.R
\name{check_prepare_default_args}
\alias{check_prepare_default_args}
\alias{.check_args_epidemic_default}
\alias{.prepare_args_epidemic_default}
\title{Check arguments to default epidemic function}
\usage{
.check_args_epidemic_default(mod_args)

.prepare_args_epidemic_default(mod_args)
}
\value{
\code{.check_args_epidemic_default()} invisibly returns the model arguments passed
in \code{mod_args}. If the model arguments did not previously contain elements
named \code{intervention} and \code{vaccination}, these are added as dummy objects of
the corresponding classes.

\code{.prepare_args_epidemic_default()} returns a list of model arguments suitable
for \code{\link[=.epidemic_default_cpp]{.epidemic_default_cpp()}}. This is a named list consisting of:
\enumerate{
\item \code{population}, the \verb{<population>} object with the contact matrix
scaled by the largest real eigenvalue and by the size of each groups; the
initial conditions are also modified to represent absolute rather than
proportional values.
\item \code{beta}, a single number for the transmission rate of the infection.
\item \code{alpha}, a single number for the rate of transition from exposed to
infectious.
\item \code{gamma}, a single number for the recovery rate.
\item \code{intervention}, the \verb{<intervention>} object,
\item \code{vaccination}, the \verb{<vaccination>} object,
\item \code{time_end}, a single number for the time point at which to end the
simulation, and
\item \code{increment},  a single number for the value by which the simulation time
is incremented.
}
}
\description{
Check and prepare the four main arguments to
\code{\link[=epidemic_default_cpp]{epidemic_default_cpp()}} for use with \code{\link[=.epidemic_default_cpp]{.epidemic_default_cpp()}}.

\code{.check_args_epidemic_default()} adds an empty \verb{<intervention>} and
\verb{<vaccination>} object if these are missing from the model arguments.

\code{.prepare_args_epidemics_default()} prepares arguments for
\code{\link[=.epidemic_default_cpp]{.epidemic_default_cpp()}}, which is the C++ function that solves the default
ODE system using a Boost \emph{odeint} solver.
\code{.prepare_args_epidemics_default()} converts the arguments collected in
\code{mod_args} into simpler structures such as lists and numeric or integer
vectors that can be interpreted as C++ types such as \code{Rcpp::List},
\code{Rcpp::NumericVector}, or \code{Eigen::MatrixXd}.
}
\keyword{internal}
