% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_vacamole.R
\name{check_prepare_vacamole_args}
\alias{check_prepare_vacamole_args}
\alias{.check_args_epidemic_vacamole}
\alias{.prepare_args_epidemic_vacamole}
\title{Check arguments to the Vacamole epidemic function}
\usage{
.check_args_epidemic_vacamole(mod_args)

.prepare_args_epidemic_vacamole(mod_args)
}
\value{
\code{.check_args_epidemic_vacamole()} invisibly returns the model arguments
passed in \code{mod_args}. If the model arguments did not previously contain
elements named \code{intervention} this is added as dummy objects of the
corresponding classes.

\code{.prepare_args_epidemic_vacamole()} returns a list of model arguments
suitable for \code{\link[=.model_vacamole_cpp]{.model_vacamole_cpp()}}. This is a named list consisting of:
\itemize{
\item \code{initial_state}: the initial conditions modified to represent absolute
rather than proportional values;
\item \code{beta}, \code{beta_v}: two numbers representing the transmission rate
of the infection for unvaccinated or single-dose vaccinated, and two-dose
vaccinated individuals, respectively;
\item \code{alpha}: a single number for the transition rate from the 'exposed' and
'exposed_vaccinated' to the 'infectious' and 'infectious_vaccinated'
compartments;
\item \code{omega}, \code{omega_v}: two numbers representing the mortality rate
of the infection for unvaccinated or single-dose vaccinated, and two-dose
vaccinated individuals, respectively;
\item \code{eta}, \code{eta_v}: two numbers representing the hospitalisation rate
of the infection for unvaccinated or single-dose vaccinated, and two-dose
vaccinated individuals, respectively;
\item \code{gamma}: a single number for the recovery rate from the infection;
\item \code{contact_matrix}, a numeric matrix for the population contact matrix
scaled by the largest real eigenvalue and by the size of each groups;
\item \code{npi_time_begin}, \code{npi_time_end}: two vectors for the start and end times
of any interventions applied;
\item \code{npi_cr}: a matrix for the age- and intervention-specific effect on social
contacts;
\item \code{vax_time_begin},\code{vax_time_end}, \code{vax_nu}: three numeric matrices for the
age- and dose-specific start times, end times, and rates of any vaccination
doses implemented;
\item \code{time_end}, \code{increment}: two numbers for the time at which to end the
simulation, and the value by which the simulation time
is incremented.
}
}
\description{
Check and prepare the four main arguments to
\code{\link[=model_vacamole_cpp]{model_vacamole_cpp()}} for use with \code{\link[=.model_vacamole_cpp]{.model_vacamole_cpp()}}.

\code{.check_args_epidemic_vacamole()} adds an empty \verb{<intervention>} object if
this is missing from the model arguments.

\code{.prepare_args_epidemics_vacamole()} prepares arguments for
\code{\link[=.model_vacamole_cpp]{.model_vacamole_cpp()}}, which is the C++ function that solves the default
ODE system using a Boost \emph{odeint} solver.
\code{.prepare_args_epidemics_vacamole()} converts the arguments collected in
\code{mod_args} into simpler structures such as lists and numeric or integer
vectors that can be interpreted as C++ types such as \code{Rcpp::List},
\code{Rcpp::NumericVector}, or \code{Eigen::MatrixXd}.
}
\keyword{internal}
