% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.epidemic_default_cpp}
\alias{.epidemic_default_cpp}
\title{Run an age-structured SEIR epidemic model}
\usage{
.epidemic_default_cpp(
  population,
  beta,
  alpha,
  gamma,
  intervention,
  time_end = 200,
  increment = 0.1
)
}
\arguments{
\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{beta}{The transmission rate \eqn{\beta}.}

\item{alpha}{The rate of transition from exposed to infectious \eqn{\alpha}.}

\item{gamma}{The recovery rate \eqn{\gamma}.}

\item{intervention}{A non-pharamaceutical intervention applied during the
course of the epidemic, with a start and end time, and age-specific effect
on contacts. See \code{\link[=intervention]{intervention()}}.}

\item{time_end}{The maximum time, defaults to 200.0.}

\item{increment}{The increment time, defaults to 0.1.}
}
\description{
Run an age-structured SEIR epidemic model
}
