% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.epidemic_ebola_cpp}
\alias{.epidemic_ebola_cpp}
\title{Run an SEIR model with Erlang passage times}
\usage{
.epidemic_ebola_cpp(
  population,
  beta,
  shape_E,
  rate_E,
  shape_I,
  rate_I,
  time_end
)
}
\arguments{
\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{beta}{The transmission rate \eqn{\beta}.}

\item{shape_E}{A single integer for the shape parameter of the Erlang
distribution of passage times through the exposed compartment.}

\item{rate_E}{A single double for the rate parameter of the Erlang
distribution of passage times through the exposed compartment.}

\item{shape_I}{A single integer for the shape parameter of the Erlang
distribution of passage times through the infectious compartment.}

\item{rate_I}{A single double for the rate parameter of the Erlang
distribution of passage times through the infectious compartment.}

\item{time_end}{A single integer for the maximum simulation time; this is
assumed to be in days.}
}
\value{
A list with two elements, \code{x}, and integer matrix with as many rows
as the number of timesteps (given by \code{time_end}), and four columns, one for
each compartment, susceptible, exposed, infectious, recovered, in that order
; and \code{times}, a vector of the simulation times, taken to be days.
This output is intended to be passed to \code{\link[=output_to_df]{output_to_df()}} to be converted
into a data.frame for further analysis.
}
\description{
Wrapper function for an SEIR compartmental model with Erlang
passage times based on Getz and Dougherty (2017) J. Biological Dynamics,
and developed to model the West African ebola virus disease outbreak of 2014
.
}
\keyword{internal}
