% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ebola copy.R
\name{.model_ebola_internal}
\alias{.model_ebola_internal}
\title{Internal code for the Ebola model}
\usage{
.model_ebola_internal(
  initial_state,
  erlang_subcompartments,
  transmission_rate,
  infectiousness_rate,
  removal_rate,
  prop_community,
  etu_risk,
  funeral_risk,
  intervention,
  time_dependence,
  time_end,
  replicates
)
}
\description{
Internal code for the Ebola model
}
\keyword{internal}
