% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_elements.R
\name{.no_population_change}
\alias{.no_population_change}
\title{Specify no population change during an epidemic}
\usage{
.no_population_change(population)
}
\arguments{
\item{population}{A \verb{<population>} for which the population change list
should be suitable.}
}
\description{
Specify no population change during an epidemic
}
\keyword{internal}
