% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_default.R
\name{.prepare_args_epidemic_default}
\alias{.prepare_args_epidemic_default}
\title{Prepare arguments to default epidemic function}
\usage{
.prepare_args_epidemic_default(mod_args)
}
\value{
Returns a list of model arguments suitable for
\code{\link[=.epidemic_default_cpp]{.epidemic_default_cpp()}}. This is a named list consisting of:
\enumerate{
\item \code{population}, the \verb{<population>} object with the contact matrix
scaled by the largest real eigenvalue and by the size of each groups; the
initial conditions are also modified to represent absolute rather than
proportional values.
\item \code{beta}, a single number for the transmission rate of the infection.
\item \code{alpha}, a single number for the rate of transition from exposed to
infectious.
\item \code{gamma}, a single number for the recovery rate.
\item \code{intervention}, the \verb{<intervention>} object,
\item \code{vaccination}, the \verb{<vaccination>} object,
\item \code{time_end}, a single number for the time point at which to end the
simulation, and
\item \code{increment},  a single number for the value by which the simulation time
is incremented.
}
}
\description{
Prepares arguments for \code{\link[=.epidemic_default_cpp]{.epidemic_default_cpp()}}, which is the
C++ function that solves the default ODE system using a Boost \emph{odeint} solver
.
}
\keyword{internal}
