% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_demic.R
\name{epi_demic}
\alias{epi_demic}
\title{Run an SIR epidemic model.}
\usage{
epi_demic(option = c("stochastic", "deterministic"), parameters = list())
}
\arguments{
\item{option}{A string, either "deterministic" or "stochastic", for which
kind of model to run.}

\item{parameters}{A parameter list created by the convenience function
\code{\link[=make_parameters_sir]{make_parameters_sir()}}.}
}
\value{
A data frame in tidy (long) format with the columns \code{time}, \code{S}, \code{I}
and \code{R} for the proportions of individuals infected at each time point.
}
\description{
Run an SIR epidemic model.
}
\examples{
epi_demic(
  option = "stochastic",
  parameters = make_parameters_sir(option = "stochastic")
)
epi_demic(
  option = "deterministic",
  parameters = make_parameters_sir(option = "deterministic")
)
}
