% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{epi_demic}
\alias{epi_demic}
\title{Stochastic SIR epidemic simulation}
\usage{
epi_demic(parameters, option = "sir_stochastic")
}
\arguments{
\item{parameters}{A named list of parameters to the simulation.}

\item{option}{A string giving the simulation option.}
}
\description{
Some description.
}
