% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_cpp.R
\name{epidemic_cpp}
\alias{epidemic_cpp}
\title{Model an epidemic}
\usage{
epidemic_cpp(
  model = "default",
  population,
  r0 = 1.5,
  preinfectious_period = 3,
  infectious_period = 7,
  intervention = no_intervention(population),
  vaccination = no_vaccination(population),
  time_end = 200,
  increment = 1
)
}
\arguments{
\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{r0}{The reproductive number of the infection. Must be a vector of the
same length as the number of demographic groups.}

\item{preinfectious_period}{The mean infections period. Must be a vector
of the same length as the number of demographic groups.}

\item{infectious_period}{The mean infections period. Must be a vector of the
same length as the number of demographic groups.}

\item{intervention}{A non-pharmaceutical intervention applied to the
population during the epidemic. See \code{\link[=intervention]{intervention()}}.}

\item{vaccination}{A vaccination regime followed during the
course of the epidemic, with a start and end time, and age-specific effect
on the transition of individuals from susceptible to vaccinated.
See \code{\link[=vaccination]{vaccination()}}.}

\item{time_end}{The maximum number of timesteps over which to run the model.}

\item{increment}{The size of the time increment.}
}
\value{
A \code{data.frame} with the columns "time", "compartment", "age_group",
"value". The comparments are "susceptible", "exposed", "infectious",
"recovered", and "vaccinated".
}
\description{
Simulate an epidemic using a deterministic, compartmental
epidemic model from the package's library of epidemic models.
The only option currently available is a SEIR-V model, with the compartments
"susceptible", "exposed", "infectious", "recovered", and "vaccinated".

Allows heterogeneity in social contact patterns, the demography distribution,
and key epidemiological parameters: the reproductive number \eqn{R_0}, and
the infectious period \eqn{1/\gamma}.
Also allows for group-specific initial proportions in each model compartment,
as well as group-specific vaccination start dates and vaccination rates.
}
\examples{
# create a population
uk_population <- population(
  name = "UK population",
  contact_matrix = matrix(1),
  demography_vector = 67e6,
  initial_conditions = matrix(
    c(0.9999, 0.0001, 0, 0, 0),
    nrow = 1, ncol = 5L
  )
)

# run epidemic simulation with no vaccination or intervention
epidemic_cpp(
  population = uk_population,
  r0 = rep(1.5, nrow(uk_population$contact_matrix)),
  preinfectious_period = rep(3, nrow(uk_population$contact_matrix)),
  infectious_period = rep(7, nrow(uk_population$contact_matrix)),
  time_end = 200,
  increment = 1
)
}
