% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_default.R
\name{epidemic_default}
\alias{epidemic_default}
\alias{epidemic_default_cpp}
\alias{epidemic_default_r}
\title{Model an SEIR-V epidemic with interventions}
\usage{
epidemic_default_cpp(
  population,
  infection,
  intervention = NULL,
  vaccination = NULL,
  time_dependence = NULL,
  time_end = 100,
  increment = 1
)

epidemic_default_r(
  population,
  infection,
  intervention = NULL,
  vaccination = NULL,
  time_dependence = NULL,
  time_end = 100,
  increment = 1
)
}
\arguments{
\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{infection}{An \code{infection} object created using \code{\link[=infection]{infection()}}. Must
have the basic reproductive number \eqn{R_0} of the infection, the
infectious period, and the pre-infectious period.
These are used to calculate the transmission rate \eqn{\beta}, the rate
at which individuals move from the 'exposed' to the 'infectious' compartment,
\eqn{\alpha}, and the recovery rate \eqn{\gamma}.}

\item{intervention}{An \verb{<intervention>} object representing an optional
non-pharmaceutical intervention applied to the population during the
epidemic. See \code{\link[=intervention]{intervention()}} for details on constructing interventions with
age-specific effects on social contacts, as well as for guidance on how to
concatenate multiple overlapping interventions into a single \verb{<intervention>}
object.}

\item{vaccination}{A \verb{<vaccination>} object representing an optional
vaccination regime with a single dose, followed during the course of the
epidemic, with a start and end time, and age-specific vaccination rates.
See \code{\link[=vaccination]{vaccination()}}.}

\item{time_dependence}{A named list where each name
is a model parameter (see \code{infection}), and each element is a function with
the first two arguments being the current simulation \code{time}, and \code{x}, a value
that is dependent on \code{time} (\code{x} represents a model parameter).
See \strong{Details} for more information.}

\item{time_end}{The maximum number of timesteps over which to run the model.
Taken as days, with a default value of 200 days.}

\item{increment}{The size of the time increment. Taken as days, with a
default value of 1 day.}
}
\value{
A \code{data.table} with the columns "time", "compartment", "age_group",
"value". The compartments correspond to the compartments of the model
chosen with \code{model}.
The current default model has the compartments "susceptible", "exposed",
"infectious", "recovered", and "vaccinated".
}
\description{
Simulate an epidemic using a deterministic, compartmental
epidemic model with the compartments
"susceptible", "exposed", "infectious", "recovered", and "vaccinated".
This model can accommodate heterogeneity in social contacts among demographic
groups, as well as differences in the sizes of demographic groups.

The \code{population} and an \code{infection} arguments are mandatory, while passing an
\code{intervention} and \code{vaccination} are optional and can be used to simulate
scenarios with different epidemic responses or different levels of the same
type of response.
See \strong{Details} for more information.
}
\details{
\code{epidemic_default_cpp()} is a wrapper function for \code{\link[=.epidemic_default_cpp]{.epidemic_default_cpp()}},
an internal C++ function that uses Boost \emph{odeint} solvers for an SEIR-V model
.
\code{\link[=.epidemic_default_cpp]{.epidemic_default_cpp()}} accepts arguments that
are created by processing the \code{population}, \code{infection}, \code{intervention} and
\code{vaccination} arguments to the wrapper function into simpler forms.

\code{epidemic_default_r()} is a wrapper around the internal function
\code{.ode_epidemic_default()}, which is passed to \code{\link[deSolve:lsoda]{deSolve::lsoda()}}.

Both models return equivalent results, but the C++ implementation is faster.

This model only allows for single, population-wide rates of
transition between the 'susceptible' and 'exposed' compartments, between the
'exposed' and 'infectious' compartments, and in the recovery rate.

Model rates or parameters can be made time-dependent by passing a function
which modifies the parameter based on the current ODE simulation time.
For example, a function that modifies the transmission rate \code{beta} could be
passed as \code{time_dependence = list(beta = function(time, x) x + sinpi(time))}.
This functionality may be used to model events that are expected to have some
effect on model parameters, such as seasonality or annual schedules such as
holidays.
}
