% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{epidemic_default_cpp}
\alias{epidemic_default_cpp}
\title{Run an age-structured SEIR epidemic model}
\usage{
epidemic_default_cpp(init, beta, alpha, gamma, time_end = 200, increment = 0.1)
}
\arguments{
\item{init}{The initial conditions, represented as a matrix, in which the
rows \eqn{i} represent age or demographic groups, and columns \eqn{j}
represent the proportions of each age group in the epidemiological
compartment. Compartments are arranged in order: "S", "E", "I", and "R".
Multiple (N) age groups are currently supported, thus \code{init} must be
a \eqn{N \times 4} matrix.}

\item{beta}{The transmission rate \eqn{\beta}.}

\item{alpha}{The rate of transition from exposed to infectious \eqn{\alpha}.}

\item{gamma}{The recovery rate \eqn{\gamma}.}

\item{time_end}{The maximum time, defaults to 200.0.}

\item{increment}{The increment time, defaults to 0.1.}
}
\description{
Run an age-structured SEIR epidemic model
}
