% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemics-package.R
\docType{package}
\name{epidemics-package}
\alias{epidemics}
\alias{epidemics-package}
\title{epidemics: A Library of Compartmental Epidemic Scenario Models}
\description{
Model epidemics using compartmental models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/epidemics}
  \item \url{https://epiverse-trace.github.io/epidemics/}
  \item Report bugs at \url{https://github.com/epiverse-trace/epidemics/issues}
}

}
\author{
\strong{Maintainer}: Pratik Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Rosalind Eggo \email{rosalind.eggo@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-0362-6717}{ORCID}) [copyright holder]
  \item Edwin Van Leeuwen \email{edwin.vanleeuwen@ukhsa.gov.uk} (\href{https://orcid.org/0000-0002-2383-5305}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Tim Taylor \email{tim.taylor@hiddenelephants.co.uk} (\href{https://orcid.org/0000-0002-8587-7113}{ORCID}) [contributor, reviewer]
  \item Hugo Gruson \email{hugo.gruson@data.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [reviewer]
  \item Joshua W. Lambert \email{joshua.lambert@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID}) [reviewer]
}

}
\keyword{internal}
