% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{intervention}
\alias{intervention}
\title{Construct a new intervention for an epidemic model}
\usage{
intervention(name = NA_character_, time_begin, time_end, contact_reduction)
}
\arguments{
\item{name}{String for the name of the intervention.}

\item{time_begin}{Single number for the start time of the intervention.}

\item{time_end}{Single number for the end time of the intervention.}

\item{contact_reduction}{A vector of the same length as the
number of demographic groups in the target population, which gives the
overall or group-specific proportion reduction in contacts respectively.}
}
\value{
An object of the \code{intervention} S3 class.
}
\description{
Construct a new intervention for an epidemic model
}
\examples{
# assuming a population with two age groups, 0 -- 18, and 18+
# an example in which schools are closed for 30 days (or other time units)
close_schools <- intervention(
  name = "close schools",
  time_begin = 50,
  time_end = 80,
  contact_reduction = c(0.5, 0.01) # reduces contacts differentially
)
close_schools 
}
