% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_default.R
\name{model_default}
\alias{model_default}
\alias{model_default_cpp}
\alias{model_default_r}
\title{Model an SEIR-V epidemic with interventions}
\usage{
model_default_cpp(
  population,
  transmissibility = 1.3/7,
  infectiousness_rate = 1/2,
  recovery_rate = 1/7,
  intervention = NULL,
  vaccination = NULL,
  time_dependence = NULL,
  time_end = 100,
  increment = 1
)

model_default_r(
  population,
  transmissibility = 1.3/7,
  infectiousness_rate = 1/2,
  recovery_rate = 1/7,
  intervention = NULL,
  vaccination = NULL,
  time_dependence = NULL,
  time_end = 100,
  increment = 1
)
}
\arguments{
\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{transmissibility}{A single number for the rate at which individuals
move from the susceptible to the exposed compartment upon contact with an
infectious individual. Often denoted as \eqn{\beta}, with
\eqn{\beta = R_0 / \text{infectious period}}.}

\item{infectiousness_rate}{A single number for the rate at which individuals
move from the exposed to the infectious compartment. Often denoted as
\eqn{\sigma}, with \eqn{\sigma = 1.0 / \text{pre-infectious period}}.
This value does not depend upon the number of infectious individuals in the
population.}

\item{recovery_rate}{A single number for the rate at which individuals move
from the infectious to the recovered compartment. Often denoted as
\eqn{\gamma}, with \eqn{\gamma = 1.0 / \text{infectious period}}.}

\item{intervention}{A named list of \verb{<intervention>}s representing optional
non-pharmaceutical or pharmaceutical interventions applied during the
epidemic. Only a single intervention on social contacts of the class
\verb{<contacts_intervention>} is allowed as the named element "contacts".
Multiple \verb{<rate_interventions>} on the model parameters are allowed; see
\strong{Details} for the model parameters for which interventions are supported.}

\item{vaccination}{A \verb{<vaccination>} object representing an optional
vaccination regime with a single dose, followed during the course of the
epidemic, with a start and end time, and age-specific vaccination rates.}

\item{time_dependence}{A named list of functions that modify model parameters
as a function of model time. List element names must correspond to model
parameter names. List elements must be functions of the form
\verb{function(time, x, ...)}, where \code{time} is the simulation \code{time} and \code{x}
represents a model parameter. The order of function arguments is important.}

\item{time_end}{The maximum number of timesteps over which to run the model.
Taken as days, with a default value of 200 days.}

\item{increment}{The size of the time increment. Taken as days, with a
default value of 1 day.}
}
\value{
An \verb{<epidemic>} class object with the model function name,
model data as "data", model parameters and components as "parameters", and
the package version as "hash".
}
\description{
Simulate an epidemic using a deterministic, compartmental
epidemic model with the compartments
"susceptible", "exposed", "infectious", "recovered", and "vaccinated".
This model can accommodate heterogeneity in social contacts among demographic
groups, as well as differences in the sizes of demographic groups.

The \code{population}, \code{transmissibility}, \code{infectiousness_rate}, and
\code{recovery_rate}
arguments are mandatory, while passing an \code{intervention} and \code{vaccination}
are optional and can be used to simulate scenarios with different epidemic
responses or different levels of the same type of response.
See \strong{Details} for more information.
}
\details{
\subsection{R and Rcpp implementations}{

\code{model_default_cpp()} is a wrapper function for the internal C++ function
\code{\link[=.model_default_cpp]{.model_default_cpp()}} that uses Boost \emph{odeint} solvers, while
\code{model_default_r()} is a wrapper around \code{\link[deSolve:lsoda]{deSolve::lsoda()}} which an R-only
implementation of the ODE system in \code{.ode_model_default()}.
Both models return equivalent results, but the C++ implementation is faster.
}

\subsection{Model parameters}{

This model only allows for single, population-wide rates of
transitions between compartments. The default values are:
\itemize{
\item Transmissibility (\eqn{\beta}, \code{transmissibility}): 0.186, assuming an
\eqn{R_0} = 1.3 and an infectious period of 7 days.
\item Infectiousness rate (\eqn{\sigma}, \code{infectiousness_rate}): 0.5, assuming
a pre-infectious period of 2 days.
\item Recovery rate (\eqn{\gamma}, \code{recovery_rate}): 0.143, assuming an
infectious period of 7 days.
}
}
}
\examples{
# create a population
uk_population <- population(
  name = "UK population",
  contact_matrix = matrix(1),
  demography_vector = 67e6,
  initial_conditions = matrix(
    c(0.9999, 0.0001, 0, 0, 0),
    nrow = 1, ncol = 5L
  )
)

# run epidemic simulation with no vaccination or intervention
data <- model_default_cpp(
  population = uk_population
)

# view the output
data
}
