% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vacamole.R
\name{model_vacamole}
\alias{model_vacamole}
\alias{model_vacamole_cpp}
\alias{model_vacamole_r}
\title{Model leaky, two-dose vaccination in an epidemic using Vacamole}
\usage{
model_vacamole_cpp(
  population,
  transmissibility = 1.3/7,
  infectiousness_rate = 1/2,
  hospitalisation_rate = 1/1000,
  mortality_rate = 1/1000,
  recovery_rate = 1/7,
  susc_reduction_vax = 0.2,
  hosp_reduction_vax = 0.2,
  mort_reduction_vax = 0.2,
  intervention = NULL,
  vaccination,
  time_dependence = NULL,
  time_end = 100,
  increment = 1
)

model_vacamole_r(
  population,
  transmissibility = 1.3/7,
  infectiousness_rate = 1/2,
  hospitalisation_rate = 1/1000,
  mortality_rate = 1/1000,
  recovery_rate = 1/7,
  susc_reduction_vax = 0.2,
  hosp_reduction_vax = 0.2,
  mort_reduction_vax = 0.2,
  intervention = NULL,
  vaccination,
  time_dependence = NULL,
  time_end = 100,
  increment = 1
)
}
\arguments{
\item{population}{An object of the \code{population} class, which holds a
population contact matrix, a demography vector, and the initial conditions
of each demographic group. See \code{\link[=population]{population()}}.}

\item{transmissibility}{A single number for the rate at which individuals
move from the susceptible to the exposed compartment upon contact with an
infectious individual. Often denoted as \eqn{\beta}, with
\eqn{\beta = R_0 / \text{infectious period}}.}

\item{infectiousness_rate}{A single number for the rate at which individuals
move from the exposed to the infectious compartment. Often denoted as
\eqn{\sigma}, with \eqn{\sigma = 1.0 / \text{pre-infectious period}}.
This value does not depend upon the number of infectious individuals in the
population.}

\item{hospitalisation_rate}{A single number for the hospitalisation rate of
infectious individuals.}

\item{mortality_rate}{A single parameter for the mortality rate of
infectious or hospitalised individuals.}

\item{recovery_rate}{A single number for the rate at which individuals move
from the infectious to the recovered compartment. Often denoted as
\eqn{\gamma}, with \eqn{\gamma = 1.0 / \text{infectious period}}.}

\item{susc_reduction_vax}{A single value between 0.0 and 1.0
giving the reduction in susceptibility of infectious individuals who
have received two doses of the vaccine.}

\item{hosp_reduction_vax}{A single value between 0.0 and 1.0
giving the reduction in hospitalisation rate of infectious individuals who
have received two doses of the vaccine.}

\item{mort_reduction_vax}{A single value between 0.0 and 1.0
giving the reduction in mortality of infectious and hospitalised individuals
who have received two doses of the vaccine.}

\item{intervention}{An \verb{<intervention>} object representing an optional
non-pharmaceutical intervention applied to the population during the
epidemic. See \code{\link[=intervention]{intervention()}} for details on constructing interventions with
age-specific effects on social contacts, as well as for guidance on how to
concatenate multiple overlapping interventions into a single \verb{<intervention>}
object.}

\item{vaccination}{A \verb{<vaccination>} object representing an optional
vaccination regime with two doses followed during the course of the
epidemic, with a start and end time, and age-specific vaccination rates for
each dose. See \code{\link[=vaccination]{vaccination()}}.}

\item{time_dependence}{A named list where each name
is a model parameter (see \code{infection}), and each element is a function with
the first two arguments being the current simulation \code{time}, and \code{x}, a value
that is dependent on \code{time} (\code{x} represents a model parameter).
See \strong{Details} for more information, as well as the vignette on time-
dependence \code{vignette("time_dependence", package = "epidemics")}.}

\item{time_end}{The maximum number of timesteps over which to run the model.
Taken as days, with a default value of 200 days.}

\item{increment}{The size of the time increment. Taken as days, with a
default value of 1 day.}
}
\value{
A \code{data.table} with the columns "time", "compartment", "age_group",
"value". The compartments correspond to the compartments of the model
chosen with \code{model}.
The current default model has the compartments "susceptible",
"vaccinated_one_dose", "vaccinated_two_dose", "exposed",
"infectious", "infectious_vaccinated", "hospitalised",
"hospitalised_vaccinated", "recovered",  and "dead".
}
\description{
Simulate an epidemic using the \emph{Vacamole} model for Covid-19
developed at RIVM, the National Institute for Public Health and the
Environment in the Netherlands.
This model is aimed at estimating the impact of 'leaky' vaccination on an
epidemic. See \strong{Details} for more information.
}
\details{
This model allows for:
\enumerate{
\item A 'hospitalised' compartment along with a hospitalisation rates;
\item Two doses of vaccination, with 'leaky' protection, i.e., vaccination does
not prevent infection completely but allows for a reduction in the infection
rate, as well as reduced rates of moving into states considered more serious,
such as 'hospitalised' or 'dead'.
}
\subsection{R and Rcpp implementations}{

\code{model_vacamole_cpp()} is a wrapper function for
\code{\link[=.model_vacamole_cpp]{.model_vacamole_cpp()}}, a C++ function that uses Boost \emph{odeint} solvers
to implement the RIVM Vacamole model.

\code{model_vacamole_r()} is a wrapper around the internal function
\code{.ode_model_vacamole()}, which is passed to \code{\link[deSolve:lsoda]{deSolve::lsoda()}}.

Both models return equivalent results, but the C++ implementation is faster.

\code{.model_vacamole_cpp()} and \code{.ode_model_vacamole()} both accept
arguments that are created by processing the \code{population}, \code{infection},
\code{intervention} and \code{vaccination} arguments to the wrapper function into
simpler forms.
}

\subsection{Model parameters}{

This model only allows for single, population-wide rates of
transition between the 'susceptible' and 'exposed' compartments, between the
'exposed' and 'infectious' compartments, and in the recovery rate.

The default values are:
\itemize{
\item Transmissibility (\eqn{\beta}, \code{transmissibility}): 0.186, resulting from
an \eqn{R_0} = 1.3 and an infectious period of 7 days.
\item Infectiousness rate (\eqn{\sigma}, \code{infectiousness_rate}): 0.5, assuming
a pre-infectious period of 2 days.
\item Hospitalisation rate (\eqn{\eta}, \code{hospitalistion_rate}): 1.0 / 1000,
assuming that one in every thousand infectious individuals is hospitalised.
\item Mortality rate (\eqn{\omega}, \code{mortality_rate}): 1.0 / 1000,
assuming that one in every thousand infectious and hospitalised
individuals dies.
\item Recovery rate (\eqn{\gamma}, \code{recovery_rate}): 0.143, assuming an
infectious period of 7 days.
\item Susceptibility reduction from vaccination (\code{susc_reduction_vax}): 0.2,
assuming a 20\% reduction in susceptibility for individuals who are doubly
vaccinated.
\item Hospitalisation reduction from vaccination (\code{hosp_reduction_vax}): 0.2,
assuming a 20\% reduction in hospitalisation for individuals who are doubly
vaccinated.
\item Mortality reduction from vaccination (\code{mort_reduction_vax}): 0.2,
assuming a 20\% reduction in mortality for individuals who are doubly
vaccinated.
}
}

\subsection{Implementing time-dependent parameters}{

Model rates or parameters can be made time-dependent by passing a function
which modifies the parameter based on the current ODE simulation time.
For example, a function that increases the transmission rate
\code{transmissibility} could be passed as
\code{time_dependence = list(transmissibility = function(time, x) x * time)}.
This functionality may be used to model events that are expected to have some
effect on model parameters, such as seasonality or annual schedules such as
holidays.
}
}
\examples{
# create a population, note eleven columns for compartments
population <- population(
  contact_matrix = matrix(1),
  demography_vector = 67e6,
  initial_conditions = matrix(
    c(0.9999, 0, 0, 0, 0, 0.0001, 0, 0, 0, 0, 0),
    nrow = 1, ncol = 11L
  )
)

# create a vaccination regime
double_vax <- vaccination(
  nu = matrix(1e-3, ncol = 2, nrow = 1),
  time_begin = matrix(c(10, 30), nrow = 1),
  time_end = matrix(c(50, 80), nrow = 1)
)

# run epidemic simulation with no vaccination or intervention
data <- model_vacamole_cpp(
  population = population,
  vaccination = double_vax
)

# view some data
head(data)
}
\references{
Ainslie, K. E. C., Backer, J. A., Boer, P. T. de, Hoek, A. J. van,
Klinkenberg, D., Altes, H. K., Leung, K. Y., Melker, H. de, Miura, F., &
Wallinga, J. (2022). A scenario modelling analysis to anticipate the impact
of COVID-19 vaccination in adolescents and children on disease outcomes in
the Netherlands, summer 2021. Eurosurveillance, 27(44), 2101090.
\doi{10.2807/1560-7917.ES.2022.27.44.2101090}
}
