% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{new_infections}
\alias{new_infections}
\title{Get new infections}
\usage{
new_infections(data, compartments_from_susceptible)
}
\arguments{
\item{data}{A \code{data.table} (or \code{data.frame}) of model output, typically
the output of \code{\link[=epidemic]{epidemic()}}.}

\item{compartments_from_susceptible}{An optional argument, for a character
vector of the names of model compartments into which individuals transition
from the "susceptible" compartment, and which are not related to infection.
A common example is a compartment for "vaccinated" individuals who are no
longer susceptible, but who should also not be counted as infected.}
}
\value{
A \code{data.table} with the same columns as \code{data}, but with the
additional variable under \code{compartment}, "new_infections", resulting in
additional rows.
}
\description{
Get new infections
}
\examples{
# create a population
uk_population <- population(
  name = "UK population",
  contact_matrix = matrix(1),
  demography_vector = 67e6,
  initial_conditions = matrix(
    c(0.9999, 0.0001, 0, 0, 0),
    nrow = 1, ncol = 5L
  )
)

# create an infection
pandemic_influenza <- infection(
  r0 = 1.5, infectious_period = 7, preinfectious_period = 3
)

# run epidemic simulation with no vaccination or intervention
data <- epidemic(
  model_name = "default",
  population = uk_population,
  infection = pandemic_influenza,
  time_end = 200,
  increment = 1
)

new_infections(data)

}
