% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{new_infections}
\alias{new_infections}
\title{Get new infections over model time}
\usage{
new_infections(data, compartments_from_susceptible = NULL, by_group = TRUE)
}
\arguments{
\item{data}{A table of model output, typically
the output of \code{\link[=model_default]{model_default()}} or similar functions.}

\item{compartments_from_susceptible}{An optional argument, for a character
vector of the names of model compartments into which individuals transition
from the "susceptible" compartment, and which are not related to infection.
A common example is a compartment for "vaccinated" individuals who are no
longer susceptible, but who should also not be counted as infected.}

\item{by_group}{A logical representing whether the epidemic size should be
returned by demographic group, or whether a single population-wide value is
returned.}
}
\value{
A table with the same columns as \code{data}, but with the
additional variable under \code{compartment}, "new_infections", resulting in
additional rows.
}
\description{
Get new infections over model time
}
\examples{
# create a population
uk_population <- population(
  contact_matrix = matrix(1),
  demography_vector = 67e6,
  initial_conditions = matrix(
    c(0.9999, 0.0001, 0, 0, 0),
    nrow = 1, ncol = 5L
  )
)


# run epidemic simulation with no vaccination or intervention
data <- model_default(
  population = uk_population,
  time_end = 200,
  increment = 1
)

new_infections(data)

}
