% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{output_to_df}
\alias{output_to_df}
\title{Convert epidemic model output to a data.frame}
\usage{
output_to_df(
  l,
  compartments = c("susceptible", "exposed", "infectious", "recovered", "vaccinated")
)
}
\arguments{
\item{l}{A list, usually the output of a call to \code{\link[=.epidemic_default_cpp]{.epidemic_default_cpp()}}.}

\item{compartments}{A character vector of the compartment names.}
}
\value{
A \code{data.frame} with N+1 columns, where N is the number of
compartments, with the additional column being "time".
}
\description{
Convert epidemic model output to a data.frame
}
