% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_ebola.R
\name{prepare_ebola_args}
\alias{prepare_ebola_args}
\alias{.check_prepare_args_ebola}
\title{Prepare arguments to ebola model function}
\usage{
.check_prepare_args_ebola(mod_args)
}
\arguments{
\item{mod_args}{A named list of composable elements; must include a
\verb{<population>}.}
}
\value{
A list of model arguments suitable for \code{\link[=.model_ebola_internal]{.model_ebola_internal()}}, if all
elements of \code{mod_args} are suitable for the Ebola model.
If any of the model composable elements are not suitable, has the side-effect
of throwing errors to the user.
If any elements of \code{mod_args} are \code{NULL}, returns dummy placeholder values
for those elements.

The returned list of prepared model arguments is a named list consisting of:
\itemize{
\item \code{"initial_state"}: the initial conditions of the simulation, which is the
number of individuals in each compartment. No age stratification allowed.
\item \code{"intervention"}: either a list of \verb{<rate_intervention>} objects, or a
dummy \verb{<rate_intervention>} if there are no interventions specified by the
user.
\item \code{"time_dependence"}: the time-dependence composable element passed by the
user. Input checks on this element occur in the higher level model function.
}
}
\description{
Prepare arguments to \code{\link[=model_ebola]{model_ebola()}} for
\code{\link[=.model_ebola_internal]{.model_ebola_internal()}}.
}
\details{
\code{.check_prepare_args_ebola()} prepares arguments for
\code{\link[=.model_ebola_internal]{.model_ebola_internal()}}, by converting some of the arguments
collected in \code{mod_args} into simpler structures that are appropriate for the
internal model function.
}
\keyword{internal}
