% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{print.intervention}
\alias{print.intervention}
\title{Print a \verb{<intervention>} object}
\usage{
\method{print}{intervention}(x, ...)
}
\arguments{
\item{x}{A \verb{<intervention>} object.}

\item{...}{Other parameters passed to \code{\link[=print]{print()}}.}
}
\value{
Invisibly returns the \verb{<intervention>} object \code{x}.
Called for printing side-effects.
}
\description{
Print a \verb{<intervention>} object
}
