% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{print.population}
\alias{print.population}
\title{Print a \code{population} object}
\usage{
\method{print}{population}(x, ...)
}
\arguments{
\item{x}{A \code{population} object.}

\item{...}{Other parameters passed to \code{\link[=print]{print()}}.}
}
\value{
None. Prints output.
}
\description{
Print a \code{population} object
}
