% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{vaccination}
\alias{vaccination}
\title{Construct a new vaccination regime for an epidemic model}
\usage{
vaccination(name = NA_character_, nu, time_begin, time_end)
}
\arguments{
\item{name}{String for the name of the vaccination regime.}

\item{nu}{Matrix for the group-specific rate of vaccination, expressed as the
rate parameter \eqn{nu}. Each element of the matrix \eqn{nu_{ij}} represents
the rate of delivering vaccine dose \eqn{j} to demographic group \eqn{i}.}

\item{time_begin}{Matrix for the start time of delivering vaccination dose
\eqn{j} to demographic group \eqn{i}.
demographic group \eqn{i}.}

\item{time_end}{Matrix for the end time of delivering vaccination dose
\eqn{j} to demographic group \eqn{i}.}
}
\value{
An object of the \code{vaccination} S3 class.
}
\description{
Construct a new vaccination regime for an epidemic model
}
\examples{
# assuming a population with two age groups, children 0 -- 5, and others 5+
# an example for childhood vaccination only
childhood_vaccination <- vaccination(
  name = "childhood_vaccination",
  time_begin = c(0, 100), # assuming a simulation over 100 days
  time_end = c(100, 100),
  nu = c(0.0001, 0.0) # over 5s never vaccinated
)
childhood_vaccination
}
